Create dedicated class for guest-ride interaction
