+++ src/openrct2/entity/Guest.cpp
@@ -30,14 +30,10 @@
 #include "../network/network.h"
 #include "../object/MusicObject.h"
 #include "../peep/GuestPathfinding.h"
-#include "../peep/RideUseSystem.h"
-#include "../rct2/RCT2.h"
 #include "../ride/Ride.h"
 #include "../ride/RideData.h"
 #include "../ride/ShopItem.h"
-#include "../ride/Station.h"
 #include "../ride/Track.h"
-#include "../ride/Vehicle.h"
 #include "../scenario/Scenario.h"
 #include "../scripting/HookEngine.h"
 #include "../scripting/ScriptEngine.h"
@@ -54,6 +50,7 @@
 #include "../world/TileElementsView.h"
 #include "Peep.h"
 #include "Staff.h"
+#include "guest/GuestRideHelper.h"
 
 #include <algorithm>
 #include <functional>
@@ -61,77 +58,7 @@
 
 using namespace OpenRCT2;
 
-// Locations of the spiral slide platform that a peep walks from the entrance of the ride to the
-// entrance of the slide. Up to 4 waypoints for each 4 sides that an ride entrance can be located
-// and 4 different rotations of the ride. 4 * 4 * 4 = 64 locations.
 // clang-format off
-static constexpr const CoordsXY SpiralSlideWalkingPath[64] = {
-    {  56,   8 },
-    {   8,   8 },
-    {   8,  32 },
-    {  32,  32 },
-    {   8,   8 },
-    {   8,   8 },
-    {   8,  32 },
-    {  32,  32 },
-    {   8,  32 },
-    {   8,  32 },
-    {   8,  32 },
-    {  32,  32 },
-    {   8,  56 },
-    {   8,  32 },
-    {   8,  32 },
-    {  32,  32 },
-    {  56,  24 },
-    {  32,  24 },
-    {  32,  24 },
-    {  32,   0 },
-    {  56, -24 },
-    {  56,  24 },
-    {  32,  24 },
-    {  32,   0 },
-    {   8,  24 },
-    {  32,  24 },
-    {  32,  24 },
-    {  32,   0 },
-    {  32,  24 },
-    {  32,  24 },
-    {  32,  24 },
-    {  32,   0 },
-    {  24,   0 },
-    {  24,   0 },
-    {  24,   0 },
-    {   0,   0 },
-    {  24, -24 },
-    {  24,   0 },
-    {  24,   0 },
-    {   0,   0 },
-    { -24, -24 },
-    {  24, -24 },
-    {  24,   0 },
-    {   0,   0 },
-    {  24,  24 },
-    {  24,   0 },
-    {  24,   0 },
-    {   0,   0 },
-    {  24,   8 },
-    {   0,   8 },
-    {   0,   8 },
-    {   0,  32 },
-    {   0,   8 },
-    {   0,   8 },
-    {   0,   8 },
-    {   0,  32 },
-    { -24,   8 },
-    {   0,   8 },
-    {   0,   8 },
-    {   0,  32 },
-    { -24,  56 },
-    { -24,   8 },
-    {   0,   8 },
-    {   0,  32 },
-};
-
 /** rct2: 0x00981F4C, 0x00981F4E */
 static constexpr const CoordsXY _WatchingPositionOffsets[] = {
     {  7,  5 },
@@ -166,13 +93,6 @@ static constexpr const CoordsXY _WatchingPositionOffsets[] = {
     {  0, 25 },
     { 25,  0 },
     {  0,  7 },
-};
-
-static constexpr const ride_rating NauseaMaximumThresholds[] = {
-    300,
-    600,
-    800,
-    1000,
 };
 
 /** rct2: 009823AC */
@@ -414,32 +334,7 @@ static struct
     { PeepActionType::Walking, PEEP_THOUGHT_ACTION_FLAG_RIDE },
 };
 
-// These arrays contain the base minimum and maximum nausea ratings for peeps, based on their nausea tolerance level.
-static constexpr const ride_rating NauseaMinimumThresholds[] = {
-    0,
-    0,
-    200,
-    400,
-};
-
-static bool peep_has_voucher_for_free_ride(Guest* peep, const Ride& ride);
-static void peep_ride_is_too_intense(Guest* peep, Ride& ride, bool peepAtRide);
-static void peep_reset_ride_heading(Guest* peep);
-static void peep_tried_to_enter_full_queue(Guest* peep, Ride& ride);
-static int16_t peep_calculate_ride_satisfaction(Guest* peep, const Ride& ride);
-static void peep_update_favourite_ride(Guest* peep, const Ride& ride);
-static int16_t peep_calculate_ride_value_satisfaction(Guest* peep, const Ride& ride);
-static int16_t peep_calculate_ride_intensity_nausea_satisfaction(Guest* peep, const Ride& ride);
-static void peep_update_ride_nausea_growth(Guest* peep, const Ride& ride);
-static bool peep_should_go_on_ride_again(Guest* peep, const Ride& ride);
-static bool peep_should_preferred_intensity_increase(Guest* peep);
-static bool peep_really_liked_ride(Guest* peep, const Ride& ride);
 static PeepThoughtType peep_assess_surroundings(int16_t centre_x, int16_t centre_y, int16_t centre_z);
-static void peep_update_hunger(Guest* peep);
-static void peep_decide_whether_to_leave_park(Guest* peep);
-static void peep_leave_park(Guest* peep);
-static void PeepHeadForNearestRideWithFlags(Guest* peep, bool considerOnlyCloseRides, int64_t rideTypeFlags);
-bool loc_690FD0(Peep* peep, RideId* rideToView, uint8_t* rideSeatToView, TileElement* tileElement);
 
 template<> bool EntityBase::Is<Guest>() const
 {
@@ -1008,13 +903,13 @@ void Guest::Tick128UpdateGuest(int32_t index)
 
             if (time_duration >= 5)
             {
-                PickRideToGoOn();
+                GuestRideHelper(*this).PickRideToGoOn();
 
                 if (GuestHeadingToRideId.IsNull())
                 {
                     HappinessTarget = std::max(HappinessTarget - 128, 0);
-                    peep_leave_park(this);
-                    peep_update_hunger(this);
+                    LeavePark();
+                    UpdateHunger();
                     loc_68F9F3();
                     loc_68FA89();
                     return;
@@ -1024,7 +919,7 @@ void Guest::Tick128UpdateGuest(int32_t index)
 
         if ((scenario_rand() & 0xFFFF) <= ((HasItem(ShopItem::Map)) ? 8192u : 2184u))
         {
-            PickRideToGoOn();
+            GuestRideHelper(*this).PickRideToGoOn();
         }
 
         if (static_cast<uint32_t>(index & 0x3FF) == (gCurrentTicks & 0x3FF))
@@ -1088,16 +983,16 @@ void Guest::Tick128UpdateGuest(int32_t index)
                     switch (chosen_thought)
                     {
                         case PeepThoughtType::Hungry:
-                            PeepHeadForNearestRideWithFlags(this, false, RIDE_TYPE_FLAG_SELLS_FOOD);
+                            GuestRideHelper(*this).HeadForNearestRideWithFlags(false, RIDE_TYPE_FLAG_SELLS_FOOD);
                             break;
                         case PeepThoughtType::Thirsty:
-                            PeepHeadForNearestRideWithFlags(this, false, RIDE_TYPE_FLAG_SELLS_DRINKS);
+                            GuestRideHelper(*this).HeadForNearestRideWithFlags(false, RIDE_TYPE_FLAG_SELLS_DRINKS);
                             break;
                         case PeepThoughtType::Toilet:
-                            PeepHeadForNearestRideWithFlags(this, false, RIDE_TYPE_FLAG_IS_TOILET);
+                            GuestRideHelper(*this).HeadForNearestRideWithFlags(false, RIDE_TYPE_FLAG_IS_TOILET);
                             break;
                         case PeepThoughtType::RunningOut:
-                            PeepHeadForNearestRideWithFlags(this, false, RIDE_TYPE_FLAG_IS_CASH_MACHINE);
+                            GuestRideHelper(*this).HeadForNearestRideWithFlags(false, RIDE_TYPE_FLAG_IS_CASH_MACHINE);
                             break;
                         default:
                             break;
@@ -1117,7 +1012,7 @@ void Guest::Tick128UpdateGuest(int32_t index)
                 if (Nausea >= 200)
                 {
                     thought_type = PeepThoughtType::VerySick;
-                    PeepHeadForNearestRideWithFlags(this, true, RIDE_TYPE_FLAG_IS_FIRST_AID);
+                    GuestRideHelper(*this).HeadForNearestRideWithFlags(true, RIDE_TYPE_FLAG_IS_FIRST_AID);
                 }
                 InsertNewThought(thought_type);
             }
@@ -1128,8 +1023,8 @@ void Guest::Tick128UpdateGuest(int32_t index)
             case PeepState::Walking:
             case PeepState::LeavingPark:
             case PeepState::EnteringPark:
-                peep_decide_whether_to_leave_park(this);
-                peep_update_hunger(this);
+                DecideWhetherToLeavePark();
+                UpdateHunger();
                 break;
 
             case PeepState::Sitting:
@@ -1148,7 +1043,7 @@ void Guest::Tick128UpdateGuest(int32_t index)
                 // In the original this branched differently
                 // but it would mean setting the peep happiness from
                 // a thought type entry which i think is incorrect.
-                peep_update_hunger(this);
+                UpdateHunger();
                 break;
 
             case PeepState::Queuing:
@@ -1193,14 +1088,14 @@ void Guest::Tick128UpdateGuest(int32_t index)
                         HappinessTarget = std::max(HappinessTarget - 4, 0);
                     }
                 }
-                peep_update_hunger(this);
+                UpdateHunger();
                 break;
             case PeepState::EnteringRide:
                 if (SubState == 17 || SubState == 15)
                 {
-                    peep_decide_whether_to_leave_park(this);
+                    DecideWhetherToLeavePark();
                 }
-                peep_update_hunger(this);
+                UpdateHunger();
                 break;
             default:
                 break;
@@ -1705,573 +1600,6 @@ bool Guest::DecideAndBuyItem(Ride& ride, ShopItem shopItem, money32 price)
     return true;
 }
 
-/**
- * Updates various peep stats upon entering a ride, as well as updating the
- * ride's satisfaction value.
- *  rct2: 0x0069545B
- */
-void Guest::OnEnterRide(Ride& ride)
-{
-    // Calculate how satisfying the ride is for the peep. Can range from -140 to +105.
-    int16_t satisfaction = peep_calculate_ride_satisfaction(this, ride);
-
-    // Update the satisfaction stat of the ride.
-    uint8_t rideSatisfaction = 0;
-    if (satisfaction >= 40)
-        rideSatisfaction = 3;
-    else if (satisfaction >= 20)
-        rideSatisfaction = 2;
-    else if (satisfaction >= 0)
-        rideSatisfaction = 1;
-
-    ride.UpdateSatisfaction(rideSatisfaction);
-
-    // Update various peep stats.
-    if (GuestNumRides < 255)
-        GuestNumRides++;
-
-    SetHasRidden(ride);
-    peep_update_favourite_ride(this, ride);
-    HappinessTarget = std::clamp(HappinessTarget + satisfaction, 0, PEEP_MAX_HAPPINESS);
-    peep_update_ride_nausea_growth(this, ride);
-}
-
-/**
- *
- *  rct2: 0x0069576E
- */
-void Guest::OnExitRide(Ride& ride)
-{
-    if (PeepFlags & PEEP_FLAGS_RIDE_SHOULD_BE_MARKED_AS_FAVOURITE)
-    {
-        PeepFlags &= ~PEEP_FLAGS_RIDE_SHOULD_BE_MARKED_AS_FAVOURITE;
-        FavouriteRide = ride.id;
-        // TODO fix this flag name or add another one
-        WindowInvalidateFlags |= PEEP_INVALIDATE_STAFF_STATS;
-    }
-    Happiness = HappinessTarget;
-    Nausea = NauseaTarget;
-    WindowInvalidateFlags |= PEEP_INVALIDATE_PEEP_STATS;
-
-    if (PeepFlags & PEEP_FLAGS_LEAVING_PARK)
-        PeepFlags &= ~(PEEP_FLAGS_PARK_ENTRANCE_CHOSEN);
-
-    if (peep_should_go_on_ride_again(this, ride))
-    {
-        GuestHeadingToRideId = ride.id;
-        GuestIsLostCountdown = 200;
-        ResetPathfindGoal();
-        WindowInvalidateFlags |= PEEP_INVALIDATE_PEEP_ACTION;
-    }
-
-    if (peep_should_preferred_intensity_increase(this))
-    {
-        if (Intensity.GetMaximum() < 15)
-        {
-            Intensity = Intensity.WithMaximum(Intensity.GetMaximum() + 1);
-        }
-    }
-
-    if (peep_really_liked_ride(this, ride))
-    {
-        InsertNewThought(PeepThoughtType::WasGreat, ride.id);
-
-        static constexpr OpenRCT2::Audio::SoundId laughs[3] = {
-            OpenRCT2::Audio::SoundId::Laugh1,
-            OpenRCT2::Audio::SoundId::Laugh2,
-            OpenRCT2::Audio::SoundId::Laugh3,
-        };
-        int32_t laughType = scenario_rand() & 7;
-        if (laughType < 3)
-        {
-            OpenRCT2::Audio::Play3D(laughs[laughType], GetLocation());
-        }
-    }
-
-    ride.total_customers++;
-    ride.window_invalidate_flags |= RIDE_INVALIDATE_RIDE_CUSTOMER;
-}
-
-/**
- *
- *  rct2: 0x00695DD2
- */
-void Guest::PickRideToGoOn()
-{
-    if (State != PeepState::Walking)
-        return;
-    if (!GuestHeadingToRideId.IsNull())
-        return;
-    if (PeepFlags & PEEP_FLAGS_LEAVING_PARK)
-        return;
-    if (HasFoodOrDrink())
-        return;
-    if (x == LOCATION_NULL)
-        return;
-
-    auto ride = FindBestRideToGoOn();
-    if (ride != nullptr)
-    {
-        // Head to that ride
-        GuestHeadingToRideId = ride->id;
-        GuestIsLostCountdown = 200;
-        ResetPathfindGoal();
-        WindowInvalidateFlags |= PEEP_INVALIDATE_PEEP_ACTION;
-
-        // Make peep look at their map if they have one
-        if (HasItem(ShopItem::Map))
-        {
-            ReadMap();
-        }
-    }
-}
-
-Ride* Guest::FindBestRideToGoOn()
-{
-    // Pick the most exciting ride
-    auto rideConsideration = FindRidesToGoOn();
-    Ride* mostExcitingRide = nullptr;
-    for (auto& ride : GetRideManager())
-    {
-        const auto rideIndex = ride.id.ToUnderlying();
-        if (rideConsideration.size() > rideIndex && rideConsideration[rideIndex])
-        {
-            if (!(ride.lifecycle_flags & RIDE_LIFECYCLE_QUEUE_FULL))
-            {
-                if (ShouldGoOnRide(ride, StationIndex::FromUnderlying(0), false, true) && ride_has_ratings(ride))
-                {
-                    if (mostExcitingRide == nullptr || ride.excitement > mostExcitingRide->excitement)
-                    {
-                        mostExcitingRide = &ride;
-                    }
-                }
-            }
-        }
-    }
-    return mostExcitingRide;
-}
-
-OpenRCT2::BitSet<OpenRCT2::Limits::MaxRidesInPark> Guest::FindRidesToGoOn()
-{
-    OpenRCT2::BitSet<OpenRCT2::Limits::MaxRidesInPark> rideConsideration;
-
-    // FIX  Originally checked for a toy, likely a mistake and should be a map,
-    //      but then again this seems to only allow the peep to go on
-    //      rides they haven't been on before.
-    if (HasItem(ShopItem::Map))
-    {
-        // Consider rides that peep hasn't been on yet
-        for (auto& ride : GetRideManager())
-        {
-            if (!HasRidden(ride))
-            {
-                rideConsideration[ride.id.ToUnderlying()] = true;
-            }
-        }
-    }
-    else
-    {
-        // Take nearby rides into consideration
-        constexpr auto radius = 10 * 32;
-        int32_t cx = floor2(x, 32);
-        int32_t cy = floor2(y, 32);
-        for (int32_t tileX = cx - radius; tileX <= cx + radius; tileX += COORDS_XY_STEP)
-        {
-            for (int32_t tileY = cy - radius; tileY <= cy + radius; tileY += COORDS_XY_STEP)
-            {
-                auto location = CoordsXY{ tileX, tileY };
-                if (!MapIsLocationValid(location))
-                    continue;
-
-                for (auto* trackElement : TileElementsView<TrackElement>(location))
-                {
-                    auto rideIndex = trackElement->GetRideIndex();
-                    if (!rideIndex.IsNull())
-                    {
-                        rideConsideration[rideIndex.ToUnderlying()] = true;
-                    }
-                }
-            }
-        }
-
-        // Always take the tall rides into consideration (realistic as you can usually see them from anywhere in the park)
-        for (auto& ride : GetRideManager())
-        {
-            if (ride.highest_drop_height > 66 || ride.excitement >= RIDE_RATING(8, 00))
-            {
-                rideConsideration[ride.id.ToUnderlying()] = true;
-            }
-        }
-    }
-
-    return rideConsideration;
-}
-
-/**
- * This function is called whenever a peep is deciding whether or not they want
- * to go on a ride or visit a shop. They may be physically present at the
- * ride/shop, or they may just be thinking about it.
- *  rct2: 0x006960AB
- */
-bool Guest::ShouldGoOnRide(Ride& ride, StationIndex entranceNum, bool atQueue, bool thinking)
-{
-    // Indicates whether a peep is physically at the ride, or is just thinking about going on the ride.
-    bool peepAtRide = !thinking;
-
-    if (ride.status == RideStatus::Open && !(ride.lifecycle_flags & RIDE_LIFECYCLE_BROKEN_DOWN))
-    {
-        // Peeps that are leaving the park will refuse to go on any rides, with the exception of free transport rides.
-        assert(ride.type < std::size(RideTypeDescriptors));
-        if (!ride.GetRideTypeDescriptor().HasFlag(RIDE_TYPE_FLAG_TRANSPORT_RIDE) || ride.value == RIDE_VALUE_UNDEFINED
-            || ride_get_price(ride) != 0)
-        {
-            if (PeepFlags & PEEP_FLAGS_LEAVING_PARK)
-            {
-                ChoseNotToGoOnRide(ride, peepAtRide, false);
-                return false;
-            }
-        }
-
-        if (ride.GetRideTypeDescriptor().HasFlag(RIDE_TYPE_FLAG_IS_SHOP_OR_FACILITY))
-        {
-            return ShouldGoToShop(ride, peepAtRide);
-        }
-
-        // This used to check !(flags & 2), but the function is only ever called with flags = 0, 1 or 6.
-        // This means we can use the existing !(flags & 4) check.
-        if (peepAtRide)
-        {
-            auto& station = ride.GetStation(entranceNum);
-
-            // Rides without queues can only have one peep waiting at a time.
-            if (!atQueue)
-            {
-                if (!station.LastPeepInQueue.IsNull())
-                {
-                    peep_tried_to_enter_full_queue(this, ride);
-                    return false;
-                }
-            }
-            else
-            {
-                // Check if there's room in the queue for the peep to enter.
-                Guest* lastPeepInQueue = GetEntity<Guest>(station.LastPeepInQueue);
-                if (lastPeepInQueue != nullptr && (abs(lastPeepInQueue->z - z) <= 6))
-                {
-                    int32_t dx = abs(lastPeepInQueue->x - x);
-                    int32_t dy = abs(lastPeepInQueue->y - y);
-                    int32_t maxD = std::max(dx, dy);
-
-                    // Unlike normal paths, peeps cannot overlap when queueing for a ride.
-                    // This check enforces a minimum distance between peeps entering the queue.
-                    if (maxD < 8)
-                    {
-                        peep_tried_to_enter_full_queue(this, ride);
-                        return false;
-                    }
-
-                    // This checks if there's a peep standing still at the very end of the queue.
-                    if (maxD <= 13 && lastPeepInQueue->TimeInQueue > 10)
-                    {
-                        peep_tried_to_enter_full_queue(this, ride);
-                        return false;
-                    }
-                }
-            }
-        }
-
-        // Assuming the queue conditions are met, peeps will always go on free transport rides.
-        // Ride ratings, recent crashes and weather will all be ignored.
-        money16 ridePrice = ride_get_price(ride);
-        if (!ride.GetRideTypeDescriptor().HasFlag(RIDE_TYPE_FLAG_TRANSPORT_RIDE) || ride.value == RIDE_VALUE_UNDEFINED
-            || ridePrice != 0)
-        {
-            if (PreviousRide == ride.id)
-            {
-                ChoseNotToGoOnRide(ride, peepAtRide, false);
-                return false;
-            }
-
-            // Basic price checks
-            if (ridePrice != 0 && !peep_has_voucher_for_free_ride(this, ride) && !(gParkFlags & PARK_FLAGS_NO_MONEY))
-            {
-                if (ridePrice > CashInPocket)
-                {
-                    if (peepAtRide)
-                    {
-                        if (CashInPocket <= 0)
-                        {
-                            InsertNewThought(PeepThoughtType::SpentMoney);
-                        }
-                        else
-                        {
-                            InsertNewThought(PeepThoughtType::CantAffordRide, ride.id);
-                        }
-                    }
-                    ChoseNotToGoOnRide(ride, peepAtRide, true);
-                    return false;
-                }
-            }
-
-            // If happy enough, peeps will ignore the fact that a ride has recently crashed.
-            if (ride.last_crash_type != RIDE_CRASH_TYPE_NONE && Happiness < 225)
-            {
-                if (peepAtRide)
-                {
-                    InsertNewThought(PeepThoughtType::NotSafe, ride.id);
-                    if (HappinessTarget >= 64)
-                    {
-                        HappinessTarget -= 8;
-                    }
-                    ride.UpdatePopularity(0);
-                }
-                ChoseNotToGoOnRide(ride, peepAtRide, true);
-                return false;
-            }
-
-            if (ride_has_ratings(ride))
-            {
-                // If a peep has already decided that they're going to go on a ride, they'll skip the weather and
-                // excitement check and will only do a basic intensity check when they arrive at the ride itself.
-                if (ride.id == GuestHeadingToRideId)
-                {
-                    if (ride.intensity > RIDE_RATING(10, 00) && !gCheatsIgnoreRideIntensity)
-                    {
-                        peep_ride_is_too_intense(this, ride, peepAtRide);
-                        return false;
-                    }
-                }
-                else
-                {
-                    if (ClimateIsRaining() && !ShouldRideWhileRaining(ride))
-                    {
-                        if (peepAtRide)
-                        {
-                            InsertNewThought(PeepThoughtType::NotWhileRaining, ride.id);
-                            if (HappinessTarget >= 64)
-                            {
-                                HappinessTarget -= 8;
-                            }
-                            ride.UpdatePopularity(0);
-                        }
-                        ChoseNotToGoOnRide(ride, peepAtRide, true);
-                        return false;
-                    }
-                    // If it is raining and the ride provides shelter skip the
-                    // ride intensity check and get me on a sheltered ride!
-                    if (!ClimateIsRaining() || !ShouldRideWhileRaining(ride))
-                    {
-                        if (!gCheatsIgnoreRideIntensity)
-                        {
-                            // Intensity calculations. Even though the max intensity can go up to 15, it's capped
-                            // at 10.0 (before happiness calculations). A full happiness bar will increase the max
-                            // intensity and decrease the min intensity by about 2.5.
-                            ride_rating maxIntensity = std::min(Intensity.GetMaximum() * 100, 1000) + Happiness;
-                            ride_rating minIntensity = (Intensity.GetMinimum() * 100) - Happiness;
-                            if (ride.intensity < minIntensity)
-                            {
-                                if (peepAtRide)
-                                {
-                                    InsertNewThought(PeepThoughtType::MoreThrilling, ride.id);
-                                    if (HappinessTarget >= 64)
-                                    {
-                                        HappinessTarget -= 8;
-                                    }
-                                    ride.UpdatePopularity(0);
-                                }
-                                ChoseNotToGoOnRide(ride, peepAtRide, true);
-                                return false;
-                            }
-                            if (ride.intensity > maxIntensity)
-                            {
-                                peep_ride_is_too_intense(this, ride, peepAtRide);
-                                return false;
-                            }
-
-                            // Nausea calculations.
-                            ride_rating maxNausea = NauseaMaximumThresholds[(EnumValue(NauseaTolerance) & 3)] + Happiness;
-
-                            if (ride.nausea > maxNausea)
-                            {
-                                if (peepAtRide)
-                                {
-                                    InsertNewThought(PeepThoughtType::Sickening, ride.id);
-                                    if (HappinessTarget >= 64)
-                                    {
-                                        HappinessTarget -= 8;
-                                    }
-                                    ride.UpdatePopularity(0);
-                                }
-                                ChoseNotToGoOnRide(ride, peepAtRide, true);
-                                return false;
-                            }
-
-                            // Very nauseous peeps will only go on very gentle rides.
-                            if (ride.nausea >= FIXED_2DP(1, 40) && Nausea > 160)
-                            {
-                                ChoseNotToGoOnRide(ride, peepAtRide, false);
-                                return false;
-                            }
-                        }
-                    }
-                }
-            }
-
-            // If the ride has not yet been rated and is capable of having g-forces,
-            // there's a 90% chance that the peep will ignore it.
-            if (!ride_has_ratings(ride) && ride.GetRideTypeDescriptor().HasFlag(RIDE_TYPE_FLAG_PEEP_CHECK_GFORCES))
-            {
-                if ((scenario_rand() & 0xFFFF) > 0x1999U)
-                {
-                    ChoseNotToGoOnRide(ride, peepAtRide, false);
-                    return false;
-                }
-
-                if (!gCheatsIgnoreRideIntensity)
-                {
-                    if (ride.max_positive_vertical_g > FIXED_2DP(5, 00) || ride.max_negative_vertical_g < FIXED_2DP(-4, 00)
-                        || ride.max_lateral_g > FIXED_2DP(4, 00))
-                    {
-                        ChoseNotToGoOnRide(ride, peepAtRide, false);
-                        return false;
-                    }
-                }
-            }
-
-            uint32_t value = ride.value;
-
-            // If the value of the ride hasn't yet been calculated, peeps will be willing to pay any amount for the ride.
-            if (value != 0xFFFF && !peep_has_voucher_for_free_ride(this, ride) && !(gParkFlags & PARK_FLAGS_NO_MONEY))
-            {
-                // The amount peeps are willing to pay is decreased by 75% if they had to pay to enter the park.
-                if (PeepFlags & PEEP_FLAGS_HAS_PAID_FOR_PARK_ENTRY)
-                    value /= 4;
-
-                // Peeps won't pay more than twice the value of the ride.
-                ridePrice = ride_get_price(ride);
-                if (ridePrice > static_cast<money16>(value * 2))
-                {
-                    if (peepAtRide)
-                    {
-                        InsertNewThought(PeepThoughtType::BadValue, ride.id);
-                        if (HappinessTarget >= 60)
-                        {
-                            HappinessTarget -= 16;
-                        }
-                        ride.UpdatePopularity(0);
-                    }
-                    ChoseNotToGoOnRide(ride, peepAtRide, true);
-                    return false;
-                }
-
-                // A ride is good value if the price is 50% or less of the ride value and the peep didn't pay to enter the park.
-                if (ridePrice <= static_cast<money16>(value / 2) && peepAtRide)
-                {
-                    if (!(gParkFlags & PARK_FLAGS_NO_MONEY))
-                    {
-                        if (!(PeepFlags & PEEP_FLAGS_HAS_PAID_FOR_PARK_ENTRY))
-                        {
-                            InsertNewThought(PeepThoughtType::GoodValue, ride.id);
-                        }
-                    }
-                }
-            }
-        }
-
-        // At this point, the peep has decided to go on the ride.
-        if (peepAtRide)
-        {
-            ride.UpdatePopularity(1);
-        }
-
-        if (ride.id == GuestHeadingToRideId)
-        {
-            peep_reset_ride_heading(this);
-        }
-
-        ride.lifecycle_flags &= ~RIDE_LIFECYCLE_QUEUE_FULL;
-        return true;
-    }
-
-    ChoseNotToGoOnRide(ride, peepAtRide, false);
-    return false;
-}
-
-bool Guest::ShouldGoToShop(Ride& ride, bool peepAtShop)
-{
-    // Peeps won't go to the same shop twice in a row.
-    if (ride.id == PreviousRide)
-    {
-        ChoseNotToGoOnRide(ride, peepAtShop, true);
-        return false;
-    }
-
-    const auto& rtd = ride.GetRideTypeDescriptor();
-    if (rtd.HasFlag(RIDE_TYPE_FLAG_IS_TOILET))
-    {
-        if (Toilet < 70)
-        {
-            ChoseNotToGoOnRide(ride, peepAtShop, true);
-            return false;
-        }
-
-        // The amount that peeps are willing to pay to use the Toilets scales with their toilet stat.
-        // It effectively has a minimum of $0.10 (due to the check above) and a maximum of $0.60.
-        if (ride_get_price(ride) * 40 > Toilet)
-        {
-            if (peepAtShop)
-            {
-                InsertNewThought(PeepThoughtType::NotPaying, ride.id);
-                if (HappinessTarget >= 60)
-                {
-                    HappinessTarget -= 16;
-                }
-                ride.UpdatePopularity(0);
-            }
-            ChoseNotToGoOnRide(ride, peepAtShop, true);
-            return false;
-        }
-    }
-
-    if (rtd.HasFlag(RIDE_TYPE_FLAG_IS_FIRST_AID))
-    {
-        if (Nausea < 128)
-        {
-            ChoseNotToGoOnRide(ride, peepAtShop, true);
-            return false;
-        }
-    }
-
-    // Basic price checks
-    auto ridePrice = ride_get_price(ride);
-    if (ridePrice != 0 && ridePrice > CashInPocket)
-    {
-        if (peepAtShop)
-        {
-            if (CashInPocket <= 0)
-            {
-                InsertNewThought(PeepThoughtType::SpentMoney);
-            }
-            else
-            {
-                InsertNewThought(PeepThoughtType::CantAffordRide, ride.id);
-            }
-        }
-        ChoseNotToGoOnRide(ride, peepAtShop, true);
-        return false;
-    }
-
-    if (peepAtShop)
-    {
-        ride.UpdatePopularity(1);
-        if (ride.id == GuestHeadingToRideId)
-        {
-            peep_reset_ride_heading(this);
-        }
-    }
-    return true;
-}
-
 // Used when no logging to an expend type required
 void Guest::SpendMoney(money32 amount, ExpenditureType expenditure)
 {
@@ -2310,28 +1638,6 @@ void Guest::SpendMoney(money16& peep_expend_type, money32 amount, ExpenditureTyp
     OpenRCT2::Audio::Play3D(OpenRCT2::Audio::SoundId::Purchase, GetLocation());
 }
 
-void Guest::SetHasRidden(const Ride& ride)
-{
-    OpenRCT2::RideUse::GetHistory().Add(sprite_index, ride.id);
-
-    SetHasRiddenRideType(ride.type);
-}
-
-bool Guest::HasRidden(const Ride& ride) const
-{
-    return OpenRCT2::RideUse::GetHistory().Contains(sprite_index, ride.id);
-}
-
-void Guest::SetHasRiddenRideType(int32_t rideType)
-{
-    OpenRCT2::RideUse::GetTypeHistory().Add(sprite_index, rideType);
-}
-
-bool Guest::HasRiddenRideType(int32_t rideType) const
-{
-    return OpenRCT2::RideUse::GetTypeHistory().Contains(sprite_index, rideType);
-}
-
 void Guest::SetParkEntryTime(int32_t entryTime)
 {
     ParkEntryTime = entryTime;
@@ -2340,39 +1646,6 @@ void Guest::SetParkEntryTime(int32_t entryTime)
 int32_t Guest::GetParkEntryTime() const
 {
     return ParkEntryTime;
-}
-
-bool Guest::ShouldRideWhileRaining(const Ride& ride)
-{
-    // Peeps will go on rides that are sufficiently undercover while it's raining.
-    // The threshold is fairly low and only requires about 10-15% of the ride to be undercover.
-    if (ride.sheltered_eighths >= 3)
-    {
-        return true;
-    }
-
-    // Peeps with umbrellas will go on rides where they can use their umbrella on it (like the Maze) 50% of the time
-    if (HasItem(ShopItem::Umbrella) && ride.GetRideTypeDescriptor().HasFlag(RIDE_TYPE_FLAG_PEEP_CAN_USE_UMBRELLA)
-        && (scenario_rand() & 2) == 0)
-    {
-        return true;
-    }
-
-    return false;
-}
-
-void Guest::ChoseNotToGoOnRide(const Ride& ride, bool peepAtRide, bool updateLastRide)
-{
-    if (peepAtRide && updateLastRide)
-    {
-        PreviousRide = ride.id;
-        PreviousRideTimeOut = 0;
-    }
-
-    if (ride.id == GuestHeadingToRideId)
-    {
-        peep_reset_ride_heading(this);
-    }
 }
 
 void Guest::ReadMap()
@@ -2384,518 +1657,6 @@ void Guest::ReadMap()
         ActionSpriteImageOffset = 0;
         UpdateCurrentActionSpriteType();
     }
-}
-
-static bool peep_has_voucher_for_free_ride(Guest* peep, const Ride& ride)
-{
-    return peep->HasItem(ShopItem::Voucher) && peep->VoucherType == VOUCHER_TYPE_RIDE_FREE && peep->VoucherRideId == ride.id;
-}
-
-/**
- * When the queue is full, peeps will ignore the ride when thinking about what to go on next.
- * Does not effect peeps that walk up to the queue entrance.
- * This flag is reset the next time a peep successfully joins the queue.
- */
-static void peep_tried_to_enter_full_queue(Guest* peep, Ride& ride)
-{
-    ride.lifecycle_flags |= RIDE_LIFECYCLE_QUEUE_FULL;
-    peep->PreviousRide = ride.id;
-    peep->PreviousRideTimeOut = 0;
-    // Change status "Heading to" to "Walking" if queue is full
-    if (ride.id == peep->GuestHeadingToRideId)
-    {
-        peep_reset_ride_heading(peep);
-    }
-}
-
-static void peep_reset_ride_heading(Guest* peep)
-{
-    peep->GuestHeadingToRideId = RideId::GetNull();
-    peep->WindowInvalidateFlags |= PEEP_INVALIDATE_PEEP_ACTION;
-}
-
-static void peep_ride_is_too_intense(Guest* peep, Ride& ride, bool peepAtRide)
-{
-    if (peepAtRide)
-    {
-        peep->InsertNewThought(PeepThoughtType::Intense, ride.id);
-        if (peep->HappinessTarget >= 64)
-        {
-            peep->HappinessTarget -= 8;
-        }
-        ride.UpdatePopularity(0);
-    }
-    peep->ChoseNotToGoOnRide(ride, peepAtRide, true);
-}
-
-/**
- *
- *  rct2: 0x00691C6E
- */
-static Vehicle* peep_choose_car_from_ride(Peep* peep, const Ride& ride, std::vector<uint8_t>& car_array)
-{
-    uint8_t chosen_car = scenario_rand();
-    if (ride.GetRideTypeDescriptor().HasFlag(RIDE_TYPE_FLAG_HAS_G_FORCES) && ((chosen_car & 0xC) != 0xC))
-    {
-        chosen_car = (scenario_rand() & 1) ? 0 : static_cast<uint8_t>(car_array.size()) - 1;
-    }
-    else
-    {
-        chosen_car = (chosen_car * static_cast<uint16_t>(car_array.size())) >> 8;
-    }
-
-    peep->CurrentCar = car_array[chosen_car];
-
-    Vehicle* vehicle = GetEntity<Vehicle>(ride.vehicles[peep->CurrentTrain]);
-    if (vehicle == nullptr)
-    {
-        return nullptr;
-    }
-    return vehicle->GetCar(peep->CurrentCar);
-}
-
-/**
- *
- *  rct2: 0x00691CD1
- */
-static void peep_choose_seat_from_car(Peep* peep, const Ride& ride, Vehicle* vehicle)
-{
-    if (vehicle == nullptr)
-    {
-        return;
-    }
-    uint8_t chosen_seat = vehicle->next_free_seat;
-
-    if (ride.mode == RideMode::ForwardRotation || ride.mode == RideMode::BackwardRotation)
-    {
-        chosen_seat = (((~vehicle->Pitch + 1) >> 3) & 0xF) * 2;
-        if (vehicle->next_free_seat & 1)
-        {
-            chosen_seat++;
-        }
-    }
-    peep->CurrentSeat = chosen_seat;
-    vehicle->next_free_seat++;
-
-    vehicle->peep[peep->CurrentSeat] = peep->sprite_index;
-    vehicle->peep_tshirt_colours[peep->CurrentSeat] = peep->TshirtColour;
-}
-
-/**
- *
- *  rct2: 0x00691D27
- */
-void Guest::GoToRideEntrance(const Ride& ride)
-{
-    const auto& station = ride.GetStation(CurrentRideStation);
-    if (station.Entrance.IsNull())
-    {
-        RemoveFromQueue();
-        return;
-    }
-
-    auto location = station.Entrance.ToCoordsXYZD().ToTileCentre();
-    int16_t x_shift = DirectionOffsets[location.direction].x;
-    int16_t y_shift = DirectionOffsets[location.direction].y;
-
-    uint8_t shift_multiplier = 21;
-    rct_ride_entry* rideEntry = get_ride_entry(ride.subtype);
-    if (rideEntry != nullptr)
-    {
-        if (rideEntry->Cars[rideEntry->DefaultCar].flags & CAR_ENTRY_FLAG_MINI_GOLF
-            || rideEntry->Cars[rideEntry->DefaultCar].flags & (CAR_ENTRY_FLAG_CHAIRLIFT | CAR_ENTRY_FLAG_GO_KART))
-        {
-            shift_multiplier = 32;
-        }
-    }
-
-    x_shift *= shift_multiplier;
-    y_shift *= shift_multiplier;
-
-    location.x += x_shift;
-    location.y += y_shift;
-
-    SetDestination(location, 2);
-    SetState(PeepState::EnteringRide);
-    RideSubState = PeepRideSubState::InEntrance;
-
-    RejoinQueueTimeout = 0;
-    GuestTimeOnRide = 0;
-
-    RemoveFromQueue();
-}
-
-bool Guest::FindVehicleToEnter(const Ride& ride, std::vector<uint8_t>& car_array)
-{
-    uint8_t chosen_train = RideStation::NO_TRAIN;
-
-    if (ride.mode == RideMode::Dodgems || ride.mode == RideMode::Race)
-    {
-        if (ride.lifecycle_flags & RIDE_LIFECYCLE_PASS_STATION_NO_STOPPING)
-            return false;
-
-        for (int32_t i = 0; i < ride.NumTrains; ++i)
-        {
-            Vehicle* vehicle = GetEntity<Vehicle>(ride.vehicles[i]);
-            if (vehicle == nullptr)
-                continue;
-
-            if (vehicle->next_free_seat >= vehicle->num_seats)
-                continue;
-
-            if (vehicle->status != Vehicle::Status::WaitingForPassengers)
-                continue;
-            chosen_train = i;
-            break;
-        }
-    }
-    else
-    {
-        chosen_train = ride.GetStation(CurrentRideStation).TrainAtStation;
-    }
-    if (chosen_train >= OpenRCT2::Limits::MaxTrainsPerRide)
-    {
-        return false;
-    }
-
-    CurrentTrain = chosen_train;
-
-    int32_t i = 0;
-
-    auto vehicle_id = ride.vehicles[chosen_train];
-    for (Vehicle* vehicle = GetEntity<Vehicle>(vehicle_id); vehicle != nullptr;
-         vehicle = GetEntity<Vehicle>(vehicle->next_vehicle_on_train), ++i)
-    {
-        uint8_t num_seats = vehicle->num_seats;
-        if (vehicle->IsUsedInPairs())
-        {
-            if (vehicle->next_free_seat & 1)
-            {
-                car_array.clear();
-                car_array.push_back(i);
-                return true;
-            }
-            num_seats &= VEHICLE_SEAT_NUM_MASK;
-        }
-        if (num_seats == vehicle->next_free_seat)
-            continue;
-
-        if (ride.mode == RideMode::ForwardRotation || ride.mode == RideMode::BackwardRotation)
-        {
-            uint8_t position = (((~vehicle->Pitch + 1) >> 3) & 0xF) * 2;
-            if (!vehicle->peep[position].IsNull())
-                continue;
-        }
-        car_array.push_back(i);
-    }
-
-    return !car_array.empty();
-}
-
-static void peep_update_ride_at_entrance_try_leave(Guest* peep)
-{
-    // Destination Tolerance is zero when peep has completely
-    // entered entrance
-    if (peep->DestinationTolerance == 0)
-    {
-        peep->RemoveFromQueue();
-        peep->SetState(PeepState::Falling);
-    }
-}
-
-static bool peep_check_ride_price_at_entrance(Guest* peep, const Ride& ride, money32 ridePrice)
-{
-    if ((peep->HasItem(ShopItem::Voucher)) && peep->VoucherType == VOUCHER_TYPE_RIDE_FREE
-        && peep->VoucherRideId == peep->CurrentRide)
-        return true;
-
-    if (peep->CashInPocket <= 0 && !(gParkFlags & PARK_FLAGS_NO_MONEY))
-    {
-        peep->InsertNewThought(PeepThoughtType::SpentMoney);
-        peep_update_ride_at_entrance_try_leave(peep);
-        return false;
-    }
-
-    if (ridePrice > peep->CashInPocket)
-    {
-        // Prevent looping of same thought / animation since Destination Tolerance
-        // is only 0 exactly at entrance and will immediately change as guest
-        // tries to leave hereafter
-        if (peep->DestinationTolerance == 0)
-        {
-            peep->InsertNewThought(PeepThoughtType::CantAffordRide, peep->CurrentRide);
-        }
-        peep_update_ride_at_entrance_try_leave(peep);
-        return false;
-    }
-
-    uint16_t value = ride.value;
-    if (value != RIDE_VALUE_UNDEFINED)
-    {
-        if (value * 2 < ridePrice)
-        {
-            peep->InsertNewThought(PeepThoughtType::BadValue, peep->CurrentRide);
-            peep_update_ride_at_entrance_try_leave(peep);
-            return false;
-        }
-    }
-    return true;
-}
-
-/**
- * The satisfaction values calculated here are used to determine how happy the peep is with the ride,
- * and also affects the satisfaction stat of the ride itself. The factors that affect satisfaction include:
- * - The price of the ride compared to the ride's value
- * - How closely the intensity and nausea of the ride matches the peep's preferences
- * - How long the peep was waiting in the queue
- * - If the peep has been on the ride before, or on another ride of the same type
- */
-static int16_t peep_calculate_ride_satisfaction(Guest* peep, const Ride& ride)
-{
-    int16_t satisfaction = peep_calculate_ride_value_satisfaction(peep, ride);
-    satisfaction += peep_calculate_ride_intensity_nausea_satisfaction(peep, ride);
-
-    // Calculate satisfaction based on how long the peep has been in the queue for.
-    // (For comparison: peeps start thinking "I've been queueing for a long time" at 3500 and
-    // start leaving the queue at 4300.)
-    if (peep->TimeInQueue >= 4500)
-        satisfaction -= 35;
-    else if (peep->TimeInQueue >= 2250)
-        satisfaction -= 10;
-    else if (peep->TimeInQueue <= 750)
-        satisfaction += 10;
-
-    // Peeps get a small boost in satisfaction if they've been on a ride of the same type before,
-    // and this boost is doubled if they've already been on this particular ride.
-    if (peep->HasRiddenRideType(ride.type))
-        satisfaction += 10;
-
-    if (peep->HasRidden(*get_ride(peep->CurrentRide)))
-        satisfaction += 10;
-
-    return satisfaction;
-}
-
-/**
- * Check to see if the specified ride should become the peep's favourite.
- * For this, a "ride rating" is calculated based on the excitement of the ride and the peep's current happiness.
- * As this value cannot exceed 255, the happier the peep is, the more irrelevant the ride's excitement becomes.
- * Due to the minimum happiness requirement, an excitement rating of more than 3.8 has no further effect.
- *
- * If the ride rating is higher than any ride the peep has already been on and the happiness criteria is met,
- * the ride becomes the peep's favourite. (This doesn't happen right away, but will be updated once the peep
- * exits the ride.)
- */
-static void peep_update_favourite_ride(Guest* peep, const Ride& ride)
-{
-    peep->PeepFlags &= ~PEEP_FLAGS_RIDE_SHOULD_BE_MARKED_AS_FAVOURITE;
-    uint8_t peepRideRating = std::clamp((ride.excitement / 4) + peep->Happiness, 0, PEEP_MAX_HAPPINESS);
-    if (peepRideRating >= peep->FavouriteRideRating)
-    {
-        if (peep->Happiness >= 160 && peep->HappinessTarget >= 160)
-        {
-            peep->FavouriteRideRating = peepRideRating;
-            peep->PeepFlags |= PEEP_FLAGS_RIDE_SHOULD_BE_MARKED_AS_FAVOURITE;
-        }
-    }
-}
-
-/* rct2: 0x00695555 */
-static int16_t peep_calculate_ride_value_satisfaction(Guest* peep, const Ride& ride)
-{
-    if (gParkFlags & PARK_FLAGS_NO_MONEY)
-    {
-        return -30;
-    }
-
-    if (ride.value == RIDE_VALUE_UNDEFINED)
-    {
-        return -30;
-    }
-
-    money16 ridePrice = ride_get_price(ride);
-    if (ride.value >= ridePrice)
-    {
-        return -5;
-    }
-
-    if ((ride.value + ((ride.value * peep->Happiness) / 256)) >= ridePrice)
-    {
-        return -30;
-    }
-
-    return 0;
-}
-
-/**
- * Calculate satisfaction based on the intensity and nausea of the ride.
- * The best possible score from this section is achieved by having the intensity and nausea
- * of the ride fall exactly within the peep's preferences, but lower scores can still be achieved
- * if the peep's happiness is enough to offset it.
- */
-static int16_t peep_calculate_ride_intensity_nausea_satisfaction(Guest* peep, const Ride& ride)
-{
-    if (!ride_has_ratings(ride))
-    {
-        return 70;
-    }
-
-    uint8_t intensitySatisfaction = 3;
-    uint8_t nauseaSatisfaction = 3;
-    ride_rating maxIntensity = peep->Intensity.GetMaximum() * 100;
-    ride_rating minIntensity = peep->Intensity.GetMinimum() * 100;
-    if (minIntensity <= ride.intensity && maxIntensity >= ride.intensity)
-    {
-        intensitySatisfaction--;
-    }
-    minIntensity -= peep->Happiness * 2;
-    maxIntensity += peep->Happiness;
-    if (minIntensity <= ride.intensity && maxIntensity >= ride.intensity)
-    {
-        intensitySatisfaction--;
-    }
-    minIntensity -= peep->Happiness * 2;
-    maxIntensity += peep->Happiness;
-    if (minIntensity <= ride.intensity && maxIntensity >= ride.intensity)
-    {
-        intensitySatisfaction--;
-    }
-
-    // Although it's not shown in the interface, a peep with Average or High nausea tolerance
-    // has a minimum preferred nausea value. (For peeps with None or Low, this is set to zero.)
-    ride_rating minNausea = NauseaMinimumThresholds[(EnumValue(peep->NauseaTolerance) & 3)];
-    ride_rating maxNausea = NauseaMaximumThresholds[(EnumValue(peep->NauseaTolerance) & 3)];
-    if (minNausea <= ride.nausea && maxNausea >= ride.nausea)
-    {
-        nauseaSatisfaction--;
-    }
-    minNausea -= peep->Happiness * 2;
-    maxNausea += peep->Happiness;
-    if (minNausea <= ride.nausea && maxNausea >= ride.nausea)
-    {
-        nauseaSatisfaction--;
-    }
-    minNausea -= peep->Happiness * 2;
-    maxNausea += peep->Happiness;
-    if (minNausea <= ride.nausea && maxNausea >= ride.nausea)
-    {
-        nauseaSatisfaction--;
-    }
-
-    uint8_t highestSatisfaction = std::max(intensitySatisfaction, nauseaSatisfaction);
-    uint8_t lowestSatisfaction = std::min(intensitySatisfaction, nauseaSatisfaction);
-
-    switch (highestSatisfaction)
-    {
-        default:
-        case 0:
-            return 70;
-        case 1:
-            switch (lowestSatisfaction)
-            {
-                default:
-                case 0:
-                    return 50;
-                case 1:
-                    return 35;
-            }
-        case 2:
-            switch (lowestSatisfaction)
-            {
-                default:
-                case 0:
-                    return 35;
-                case 1:
-                    return 20;
-                case 2:
-                    return 10;
-            }
-        case 3:
-            switch (lowestSatisfaction)
-            {
-                default:
-                case 0:
-                    return -35;
-                case 1:
-                    return -50;
-                case 2:
-                    return -60;
-                case 3:
-                    return -60;
-            }
-    }
-}
-
-/**
- * Update the nausea growth of the peep based on a ride. This is calculated based on:
- * - The nausea rating of the ride
- * - Their new happiness growth rate (the higher, the less nauseous)
- * - How hungry the peep is (+0% nausea at 50% hunger up to +100% nausea at 100% hunger)
- * - The peep's nausea tolerance (Final modifier: none: 100%, low: 50%, average: 25%, high: 12.5%)
- */
-static void peep_update_ride_nausea_growth(Guest* peep, const Ride& ride)
-{
-    uint32_t nauseaMultiplier = std::clamp(256 - peep->HappinessTarget, 64, 200);
-    uint32_t nauseaGrowthRateChange = (ride.nausea * nauseaMultiplier) / 512;
-    nauseaGrowthRateChange *= std::max(static_cast<uint8_t>(128), peep->Hunger) / 64;
-    nauseaGrowthRateChange >>= (EnumValue(peep->NauseaTolerance) & 3);
-    peep->NauseaTarget = static_cast<uint8_t>(std::min(peep->NauseaTarget + nauseaGrowthRateChange, 255u));
-}
-
-static bool peep_should_go_on_ride_again(Guest* peep, const Ride& ride)
-{
-    if (!ride.GetRideTypeDescriptor().HasFlag(RIDE_TYPE_FLAG_PEEP_WILL_RIDE_AGAIN))
-        return false;
-    if (!ride_has_ratings(ride))
-        return false;
-    if (ride.intensity > RIDE_RATING(10, 00) && !gCheatsIgnoreRideIntensity)
-        return false;
-    if (peep->Happiness < 180)
-        return false;
-    if (peep->Energy < 100)
-        return false;
-    if (peep->Nausea > 160)
-        return false;
-    if (peep->Hunger < 30)
-        return false;
-    if (peep->Thirst < 20)
-        return false;
-    if (peep->Toilet > 170)
-        return false;
-
-    uint8_t r = (scenario_rand() & 0xFF);
-    if (r <= 128)
-    {
-        if (peep->GuestNumRides > 7)
-            return false;
-        if (r > 64)
-            return false;
-    }
-
-    return true;
-}
-
-static bool peep_should_preferred_intensity_increase(Guest* peep)
-{
-    if (gParkFlags & PARK_FLAGS_PREF_LESS_INTENSE_RIDES)
-        return false;
-    if (peep->Happiness < 200)
-        return false;
-
-    return (scenario_rand() & 0xFF) >= static_cast<uint8_t>(peep->Intensity);
-}
-
-static bool peep_really_liked_ride(Guest* peep, const Ride& ride)
-{
-    if (peep->Happiness < 215)
-        return false;
-    if (peep->Nausea > 120)
-        return false;
-    if (!ride_has_ratings(ride))
-        return false;
-    if (ride.intensity > RIDE_RATING(10, 00) && !gCheatsIgnoreRideIntensity)
-        return false;
-    return true;
 }
 
 /**
@@ -3021,14 +1782,14 @@ static PeepThoughtType peep_assess_surroundings(int16_t centre_x, int16_t centre
  *
  *  rct2: 0x0068F9A9
  */
-static void peep_update_hunger(Guest* peep)
+void Guest::UpdateHunger()
 {
-    if (peep->Hunger >= 3)
+    if (Hunger >= 3)
     {
-        peep->Hunger -= 2;
+        Hunger -= 2;
 
-        peep->EnergyTarget = std::min(peep->EnergyTarget + 2, PEEP_MAX_ENERGY_TARGET);
-        peep->Toilet = std::min(peep->Toilet + 1, 255);
+        EnergyTarget = std::min(EnergyTarget + 2, PEEP_MAX_ENERGY_TARGET);
+        Toilet = std::min(Toilet + 1, 255);
     }
 }
 
@@ -3038,19 +1799,19 @@ static void peep_update_hunger(Guest* peep)
  *
  *  rct2: 0x0068F8CD
  */
-static void peep_decide_whether_to_leave_park(Guest* peep)
+void Guest::DecideWhetherToLeavePark()
 {
-    if (peep->EnergyTarget >= 33)
+    if (EnergyTarget >= 33)
     {
-        peep->EnergyTarget -= 2;
+        EnergyTarget -= 2;
     }
 
-    if (gClimateCurrent.Temperature >= 21 && peep->Thirst >= 5)
+    if (gClimateCurrent.Temperature >= 21 && Thirst >= 5)
     {
-        peep->Thirst--;
+        Thirst--;
     }
 
-    if (peep->OutsideOfPark)
+    if (OutsideOfPark)
     {
         return;
     }
@@ -3058,18 +1819,18 @@ static void peep_decide_whether_to_leave_park(Guest* peep)
     /* Peeps that are happy enough, have enough energy and
      * (if appropriate) have enough money will always stay
      * in the park. */
-    if (!(peep->PeepFlags & PEEP_FLAGS_LEAVING_PARK))
+    if (!(PeepFlags & PEEP_FLAGS_LEAVING_PARK))
     {
         if (gParkFlags & PARK_FLAGS_NO_MONEY)
         {
-            if (peep->Energy >= 70 && peep->Happiness >= 60)
+            if (Energy >= 70 && Happiness >= 60)
             {
                 return;
             }
         }
         else
         {
-            if (peep->Energy >= 55 && peep->Happiness >= 45 && peep->CashInPocket >= 5.00_GBP)
+            if (Energy >= 55 && Happiness >= 45 && CashInPocket >= 5.00_GBP)
             {
                 return;
             }
@@ -3083,153 +1844,36 @@ static void peep_decide_whether_to_leave_park(Guest* peep)
     }
 
     // In the remaining 5% chance the peep leaves the park.
-    peep_leave_park(peep);
+    LeavePark();
 }
 
 /**
  *
  *  rct2: 0x0068F93E
  */
-static void peep_leave_park(Guest* peep)
+void Guest::LeavePark()
 {
-    peep->GuestHeadingToRideId = RideId::GetNull();
-    if (peep->PeepFlags & PEEP_FLAGS_LEAVING_PARK)
+    GuestHeadingToRideId = RideId::GetNull();
+    if (PeepFlags & PEEP_FLAGS_LEAVING_PARK)
     {
-        if (peep->GuestIsLostCountdown < 60)
+        if (GuestIsLostCountdown < 60)
         {
             return;
         }
     }
     else
     {
-        peep->GuestIsLostCountdown = 254;
-        peep->PeepFlags |= PEEP_FLAGS_LEAVING_PARK;
-        peep->PeepFlags &= ~PEEP_FLAGS_PARK_ENTRANCE_CHOSEN;
+        GuestIsLostCountdown = 254;
+        PeepFlags |= PEEP_FLAGS_LEAVING_PARK;
+        PeepFlags &= ~PEEP_FLAGS_PARK_ENTRANCE_CHOSEN;
     }
 
-    peep->InsertNewThought(PeepThoughtType::GoHome);
+    InsertNewThought(PeepThoughtType::GoHome);
 
-    rct_window* w = window_find_by_number(WindowClass::Peep, peep->sprite_index);
+    rct_window* w = window_find_by_number(WindowClass::Peep, sprite_index);
     if (w != nullptr)
         window_event_invalidate_call(w);
-    window_invalidate_by_number(WindowClass::Peep, peep->sprite_index);
-}
-
-template<typename T> static void peep_head_for_nearest_ride(Guest* peep, bool considerOnlyCloseRides, T predicate)
-{
-    if (peep->State != PeepState::Sitting && peep->State != PeepState::Watching && peep->State != PeepState::Walking)
-    {
-        return;
-    }
-    if (peep->PeepFlags & PEEP_FLAGS_LEAVING_PARK)
-        return;
-    if (peep->x == LOCATION_NULL)
-        return;
-    if (!peep->GuestHeadingToRideId.IsNull())
-    {
-        auto ride = get_ride(peep->GuestHeadingToRideId);
-        if (ride != nullptr && predicate(*ride))
-        {
-            return;
-        }
-    }
-
-    OpenRCT2::BitSet<OpenRCT2::Limits::MaxRidesInPark> rideConsideration;
-    if (!considerOnlyCloseRides && (peep->HasItem(ShopItem::Map)))
-    {
-        // Consider all rides in the park
-        for (const auto& ride : GetRideManager())
-        {
-            if (predicate(ride))
-            {
-                rideConsideration[ride.id.ToUnderlying()] = true;
-            }
-        }
-    }
-    else
-    {
-        // Take nearby rides into consideration
-        constexpr auto searchRadius = 10 * 32;
-        int32_t cx = floor2(peep->x, 32);
-        int32_t cy = floor2(peep->y, 32);
-        for (auto x = cx - searchRadius; x <= cx + searchRadius; x += COORDS_XY_STEP)
-        {
-            for (auto y = cy - searchRadius; y <= cy + searchRadius; y += COORDS_XY_STEP)
-            {
-                auto location = CoordsXY{ x, y };
-                if (!MapIsLocationValid(location))
-                    continue;
-
-                for (auto* trackElement : TileElementsView<TrackElement>(location))
-                {
-                    auto rideIndex = trackElement->GetRideIndex();
-                    auto ride = get_ride(rideIndex);
-                    if (ride == nullptr)
-                        continue;
-
-                    if (!predicate(*ride))
-                        continue;
-
-                    rideConsideration[ride->id.ToUnderlying()] = true;
-                }
-            }
-        }
-    }
-
-    // Filter the considered rides
-    RideId potentialRides[OpenRCT2::Limits::MaxRidesInPark];
-    size_t numPotentialRides = 0;
-    for (auto& ride : GetRideManager())
-    {
-        if (rideConsideration[ride.id.ToUnderlying()])
-        {
-            if (!(ride.lifecycle_flags & RIDE_LIFECYCLE_QUEUE_FULL))
-            {
-                if (peep->ShouldGoOnRide(ride, StationIndex::FromUnderlying(0), false, true))
-                {
-                    potentialRides[numPotentialRides++] = ride.id;
-                }
-            }
-        }
-    }
-
-    // Pick the closest ride
-    Ride* closestRide{};
-    auto closestRideDistance = std::numeric_limits<int32_t>::max();
-    for (size_t i = 0; i < numPotentialRides; i++)
-    {
-        auto ride = get_ride(potentialRides[i]);
-        if (ride != nullptr)
-        {
-            auto rideLocation = ride->GetStation().Start;
-            int32_t distance = abs(rideLocation.x - peep->x) + abs(rideLocation.y - peep->y);
-            if (distance < closestRideDistance)
-            {
-                closestRide = ride;
-                closestRideDistance = distance;
-            }
-        }
-    }
-    if (closestRide != nullptr)
-    {
-        // Head to that ride
-        peep->GuestHeadingToRideId = closestRide->id;
-        peep->GuestIsLostCountdown = 200;
-        peep->ResetPathfindGoal();
-        peep->WindowInvalidateFlags |= PEEP_INVALIDATE_PEEP_ACTION;
-        peep->TimeLost = 0;
-    }
-}
-
-static void PeepHeadForNearestRideWithFlags(Guest* peep, bool considerOnlyCloseRides, int64_t rideTypeFlags)
-{
-    if ((rideTypeFlags & RIDE_TYPE_FLAG_IS_TOILET) && peep->HasFoodOrDrink())
-    {
-        return;
-    }
-    peep_head_for_nearest_ride(peep, considerOnlyCloseRides, [rideTypeFlags](const Ride& ride) {
-        return ride.GetRideTypeDescriptor().HasFlag(rideTypeFlags);
-    });
+    window_invalidate_by_number(WindowClass::Peep, sprite_index);
 }
 
 /**
@@ -3424,1683 +2068,24 @@ void Guest::UpdateBuying()
     SubState = 1;
 }
 
-/**
- *
- *  rct2: 0x00691A3B
- */
-void Guest::UpdateRideAtEntrance()
-{
-    auto ride = get_ride(CurrentRide);
-    if (ride == nullptr)
-        return;
-
-    // The peep will keep advancing in the entranceway
-    // whilst in this state. When it has reached the very
-    // front of the queue destination tolerance is set to
-    // zero to indicate it is final decision time (try_leave will pass).
-    // When a peep has to return to the queue without getting on a ride
-    // this is the state it will return to.
-    if (DestinationTolerance != 0)
-    {
-        int16_t xy_distance;
-        if (auto loc = UpdateAction(xy_distance); loc.has_value())
-        {
-            int16_t actionZ = z;
-            if (xy_distance < 16)
-            {
-                const auto& station = ride->GetStation(CurrentRideStation);
-                auto entrance = station.Entrance.ToCoordsXYZ();
-                actionZ = entrance.z + 2;
-            }
-            MoveTo({ loc.value(), actionZ });
-        }
-        else
-        {
-            DestinationTolerance = 0;
-            sprite_direction ^= (1 << 4);
-            Invalidate();
-        }
-    }
-
-    std::vector<uint8_t> carArray;
-
-    if (ride->GetRideTypeDescriptor().HasFlag(RIDE_TYPE_FLAG_NO_VEHICLES))
-    {
-        if (ride->num_riders >= ride->operation_option)
-            return;
-    }
-    else
-    {
-        if (!FindVehicleToEnter(*ride, carArray))
-            return;
-    }
-
-    if (ride->status != RideStatus::Open || ride->vehicle_change_timeout != 0)
-    {
-        peep_update_ride_at_entrance_try_leave(this);
-        return;
-    }
-
-    if (ride->lifecycle_flags & RIDE_LIFECYCLE_BROKEN_DOWN)
-        return;
-
-    money16 ridePrice = ride_get_price(*ride);
-    if (ridePrice != 0)
-    {
-        if (!peep_check_ride_price_at_entrance(this, *ride, ridePrice))
-            return;
-    }
-
-    if (!ride->GetRideTypeDescriptor().HasFlag(RIDE_TYPE_FLAG_NO_VEHICLES))
-    {
-        Vehicle* vehicle = peep_choose_car_from_ride(this, *ride, carArray);
-        peep_choose_seat_from_car(this, *ride, vehicle);
-    }
-    GoToRideEntrance(*ride);
-}
-
-/** rct2: 0x00981FD4, 0x00981FD6 */
-static constexpr const CoordsXY _MazeEntranceStart[] = {
-    { 8, 8 },
-    { 8, 24 },
-    { 24, 24 },
-    { 24, 8 },
-};
-
-void PeepUpdateRideLeaveEntranceMaze(Guest* peep, Ride& ride, CoordsXYZD& entrance_loc)
-{
-    peep->MazeLastEdge = entrance_loc.direction + 1;
-
-    entrance_loc.x += CoordsDirectionDelta[entrance_loc.direction].x;
-    entrance_loc.y += CoordsDirectionDelta[entrance_loc.direction].y;
-
-    uint8_t direction = entrance_loc.direction * 4 + 11;
-    if (scenario_rand() & 0x40)
-    {
-        direction += 4;
-        peep->MazeLastEdge += 2;
-    }
-
-    direction &= 0xF;
-    // Direction is 11, 15, 3, or 7
-    peep->Var37 = direction;
-    peep->MazeLastEdge &= 3;
-
-    entrance_loc.x += _MazeEntranceStart[direction / 4].x;
-    entrance_loc.y += _MazeEntranceStart[direction / 4].y;
-
-    peep->SetDestination(entrance_loc, 3);
-
-    ride.cur_num_customers++;
-    peep->OnEnterRide(ride);
-    peep->RideSubState = PeepRideSubState::MazePathfinding;
-}
-
-void PeepUpdateRideLeaveEntranceSpiralSlide(Guest* peep, Ride& ride, CoordsXYZD& entrance_loc)
-{
-    entrance_loc = { ride.GetStation(peep->CurrentRideStation).GetStart(), entrance_loc.direction };
-
-    TileElement* tile_element = ride_get_station_start_track_element(ride, peep->CurrentRideStation);
-
-    uint8_t direction_track = (tile_element == nullptr ? 0 : tile_element->GetDirection());
-
-    peep->Var37 = (entrance_loc.direction << 2) | (direction_track << 4);
-
-    entrance_loc += SpiralSlideWalkingPath[peep->Var37];
-
-    peep->SetDestination(entrance_loc);
-    peep->CurrentCar = 0;
-
-    ride.cur_num_customers++;
-    peep->OnEnterRide(ride);
-    peep->RideSubState = PeepRideSubState::ApproachSpiralSlide;
-}
-
-void PeepUpdateRideLeaveEntranceDefault(Guest* peep, Ride& ride, CoordsXYZD& entrance_loc)
-{
-    // If the ride type was changed guests will become stuck.
-    // Inform the player about this if its a new issue or hasn't been addressed within 120 seconds.
-    if ((ride.current_issues & RIDE_ISSUE_GUESTS_STUCK) == 0 || gCurrentTicks - ride.last_issue_time > 3000)
-    {
-        ride.current_issues |= RIDE_ISSUE_GUESTS_STUCK;
-        ride.last_issue_time = gCurrentTicks;
-
-        auto ft = Formatter();
-        ride.FormatNameTo(ft);
-        if (gConfigNotifications.RideWarnings)
-        {
-            News::AddItemToQueue(News::ItemType::Ride, STR_GUESTS_GETTING_STUCK_ON_RIDE, peep->CurrentRide.ToUnderlying(), ft);
-        }
-    }
-}
-
-uint8_t Guest::GetWaypointedSeatLocation(const Ride& ride, CarEntry* vehicle_type, uint8_t track_direction) const
-{
-    // The seatlocation can be split into segments around the ride base
-    // to decide the segment first split off the segmentable seat location
-    // from the fixed section
-    uint8_t seatLocationSegment = CurrentSeat & 0x7;
-    uint8_t seatLocationFixed = CurrentSeat & 0xF8;
-
-    // Enterprise has more segments (8) compared to the normal (4)
-    if (ride.type != RIDE_TYPE_ENTERPRISE)
-        track_direction *= 2;
-
-    // Type 1 loading doesn't do segments and all peeps go to the same
-    // location on the ride
-    if (vehicle_type->peep_loading_waypoint_segments == 0)
-    {
-        track_direction /= 2;
-        seatLocationSegment = 0;
-        seatLocationFixed = 0;
-    }
-    seatLocationSegment += track_direction;
-    seatLocationSegment &= 0x7;
-    return seatLocationSegment + seatLocationFixed;
-}
-
-void Guest::UpdateRideLeaveEntranceWaypoints(const Ride& ride)
-{
-    const auto& station = ride.GetStation(CurrentRideStation);
-    Guard::Assert(!station.Entrance.IsNull());
-    uint8_t direction_entrance = station.Entrance.direction;
-
-    TileElement* tile_element = ride_get_station_start_track_element(ride, CurrentRideStation);
-
-    uint8_t direction_track = (tile_element == nullptr ? 0 : tile_element->GetDirection());
-
-    auto vehicle = GetEntity<Vehicle>(ride.vehicles[CurrentTrain]);
-    if (vehicle == nullptr)
-    {
-        // TODO: Goto ride exit on failure.
-        return;
-    }
-    auto ride_entry = vehicle->GetRideEntry();
-    auto vehicle_type = &ride_entry->Cars[vehicle->vehicle_type];
-
-    Var37 = (direction_entrance | GetWaypointedSeatLocation(ride, vehicle_type, direction_track) * 4) * 4;
-
-    const auto& rtd = ride.GetRideTypeDescriptor();
-    CoordsXY waypoint = rtd.GetGuestWaypointLocation(*vehicle, ride, CurrentRideStation);
-
-    const auto waypointIndex = Var37 / 4;
-    Guard::Assert(vehicle_type->peep_loading_waypoints.size() >= static_cast<size_t>(waypointIndex));
-    waypoint.x += vehicle_type->peep_loading_waypoints[waypointIndex][0].x;
-    waypoint.y += vehicle_type->peep_loading_waypoints[waypointIndex][0].y;
-
-    SetDestination(waypoint);
-    RideSubState = PeepRideSubState::ApproachVehicleWaypoints;
-}
-
-/**
- *
- *  rct2: 0x006921D3
- */
-void Guest::UpdateRideAdvanceThroughEntrance()
-{
-    auto ride = get_ride(CurrentRide);
-    if (ride == nullptr)
-        return;
-
-    int16_t actionZ, xy_distance;
-
-    auto ride_entry = ride->GetRideEntry();
-
-    if (auto loc = UpdateAction(xy_distance); loc.has_value())
-    {
-        uint16_t distanceThreshold = 16;
-        if (ride_entry != nullptr)
-        {
-            uint8_t vehicle = ride_entry->DefaultCar;
-            if (ride_entry->Cars[vehicle].flags & CAR_ENTRY_FLAG_MINI_GOLF
-                || ride_entry->Cars[vehicle].flags & (CAR_ENTRY_FLAG_CHAIRLIFT | CAR_ENTRY_FLAG_GO_KART))
-            {
-                distanceThreshold = 28;
-            }
-        }
-
-        if (RideSubState == PeepRideSubState::InEntrance && xy_distance < distanceThreshold)
-        {
-            RideSubState = PeepRideSubState::FreeVehicleCheck;
-        }
-
-        actionZ = ride->GetStation(CurrentRideStation).GetBaseZ();
-
-        distanceThreshold += 4;
-        if (xy_distance < distanceThreshold)
-        {
-            actionZ += ride->GetRideTypeDescriptor().Heights.PlatformHeight;
-        }
-
-        MoveTo({ loc.value(), actionZ });
-        return;
-    }
-
-    if (RideSubState == PeepRideSubState::InEntrance)
-    {
-        RideSubState = PeepRideSubState::FreeVehicleCheck;
-        return;
-    }
-
-    if (ride->GetRideTypeDescriptor().HasFlag(RIDE_TYPE_FLAG_NO_VEHICLES))
-    {
-        const auto& station = ride->GetStation(CurrentRideStation);
-        auto entranceLocation = station.Entrance.ToCoordsXYZD();
-        Guard::Assert(!entranceLocation.IsNull());
-
-        const auto& rtd = GetRideTypeDescriptor(ride->type);
-        rtd.UpdateLeaveEntrance(this, *ride, entranceLocation);
-        return;
-    }
-
-    Vehicle* vehicle = GetEntity<Vehicle>(ride->vehicles[CurrentTrain]);
-    if (vehicle == nullptr)
-    {
-        return;
-    }
-
-    vehicle = vehicle->GetCar(CurrentCar);
-    if (vehicle == nullptr)
-    {
-        return;
-    }
-
-    ride_entry = vehicle->GetRideEntry();
-    if (ride_entry == nullptr)
-    {
-        return;
-    }
-
-    CarEntry* vehicle_type = &ride_entry->Cars[vehicle->vehicle_type];
-
-    if (vehicle_type->flags & CAR_ENTRY_FLAG_LOADING_WAYPOINTS)
-    {
-        UpdateRideLeaveEntranceWaypoints(*ride);
-        return;
-    }
-
-    if (vehicle_type->flags & CAR_ENTRY_FLAG_DODGEM_CAR_PLACEMENT)
-    {
-        SetDestination(vehicle->GetLocation(), 15);
-        RideSubState = PeepRideSubState::ApproachVehicle;
-        return;
-    }
-
-    int8_t load_position = 0;
-    // Safe, in case current seat > number of loading positions
-    uint16_t numSeatPositions = static_cast<uint16_t>(vehicle_type->peep_loading_positions.size());
-    if (numSeatPositions != 0)
-    {
-        size_t loadPositionIndex = numSeatPositions - 1;
-        if (CurrentSeat < numSeatPositions)
-        {
-            loadPositionIndex = CurrentSeat;
-        }
-        load_position = vehicle_type->peep_loading_positions[loadPositionIndex];
-    }
-
-    auto destination = GetDestination();
-    switch (vehicle->sprite_direction / 8)
-    {
-        case 0:
-            destination.x = vehicle->x - load_position;
-            break;
-        case 1:
-            destination.y = vehicle->y + load_position;
-            break;
-        case 2:
-            destination.x = vehicle->x + load_position;
-            break;
-        case 3:
-            destination.y = vehicle->y - load_position;
-            break;
-    }
-    SetDestination(destination);
-
-    RideSubState = PeepRideSubState::ApproachVehicle;
-}
-
-/**
- *
- *  rct2: 0x0069321D
- */
-static void peep_go_to_ride_exit(Peep* peep, const Ride& ride, int16_t x, int16_t y, int16_t z, uint8_t exit_direction)
-{
-    z += ride.GetRideTypeDescriptor().Heights.PlatformHeight;
-
-    peep->MoveTo({ x, y, z });
-
-    Guard::Assert(peep->CurrentRideStation.ToUnderlying() < OpenRCT2::Limits::MaxStationsPerRide);
-    auto exit = ride.GetStation(peep->CurrentRideStation).Exit;
-    Guard::Assert(!exit.IsNull());
-    x = exit.x;
-    y = exit.y;
-    x *= 32;
-    y *= 32;
-    x += 16;
-    y += 16;
-
-    int16_t x_shift = DirectionOffsets[exit_direction].x;
-    int16_t y_shift = DirectionOffsets[exit_direction].y;
-
-    int16_t shift_multiplier = 20;
-
-    rct_ride_entry* rideEntry = get_ride_entry(ride.subtype);
-    if (rideEntry != nullptr)
-    {
-        CarEntry* carEntry = &rideEntry->Cars[rideEntry->DefaultCar];
-        if (carEntry->flags & CAR_ENTRY_FLAG_MINI_GOLF || carEntry->flags & (CAR_ENTRY_FLAG_CHAIRLIFT | CAR_ENTRY_FLAG_GO_KART))
-        {
-            shift_multiplier = 32;
-        }
-    }
-
-    x_shift *= shift_multiplier;
-    y_shift *= shift_multiplier;
-
-    x -= x_shift;
-    y -= y_shift;
-
-    peep->SetDestination({ x, y }, 2);
-
-    peep->sprite_direction = exit_direction * 8;
-    peep->RideSubState = PeepRideSubState::ApproachExit;
-}
-
-/**
- *
- *  rct2: 0x006920B4
- */
-void Guest::UpdateRideFreeVehicleEnterRide(Ride& ride)
-{
-    money16 ridePrice = ride_get_price(ride);
-    if (ridePrice != 0)
-    {
-        if ((HasItem(ShopItem::Voucher)) && (VoucherType == VOUCHER_TYPE_RIDE_FREE) && (VoucherRideId == CurrentRide))
-        {
-            RemoveItem(ShopItem::Voucher);
-            WindowInvalidateFlags |= PEEP_INVALIDATE_PEEP_INVENTORY;
-        }
-        else
-        {
-            ride.total_profit += ridePrice;
-            ride.window_invalidate_flags |= RIDE_INVALIDATE_RIDE_INCOME;
-            SpendMoney(PaidOnRides, ridePrice, ExpenditureType::ParkRideTickets);
-        }
-    }
-
-    RideSubState = PeepRideSubState::LeaveEntrance;
-    uint8_t queueTime = DaysInQueue;
-    if (queueTime < 253)
-        queueTime += 3;
-
-    queueTime /= 2;
-    auto& station = ride.GetStation(CurrentRideStation);
-    if (queueTime != station.QueueTime)
-    {
-        station.QueueTime = queueTime;
-        window_invalidate_by_number(WindowClass::Ride, CurrentRide.ToUnderlying());
-    }
-
-    if (PeepFlags & PEEP_FLAGS_TRACKING)
-    {
-        auto ft = Formatter();
-        FormatNameTo(ft);
-        ride.FormatNameTo(ft);
-
-        StringId msg_string;
-        if (ride.GetRideTypeDescriptor().HasFlag(RIDE_TYPE_FLAG_IN_RIDE))
-            msg_string = STR_PEEP_TRACKING_PEEP_IS_IN_X;
-        else
-            msg_string = STR_PEEP_TRACKING_PEEP_IS_ON_X;
-
-        if (gConfigNotifications.GuestOnRide)
-        {
-            News::AddItemToQueue(News::ItemType::PeepOnRide, msg_string, sprite_index, ft);
-        }
-    }
-
-    const auto& rtd = ride.GetRideTypeDescriptor();
-    if (rtd.HasFlag(RIDE_TYPE_FLAG_IS_SPIRAL_SLIDE))
-    {
-        SwitchToSpecialSprite(1);
-    }
-
-    UpdateRideAdvanceThroughEntrance();
-}
-
-/**
- *
- *  rct2: 0x00691FD4
- */
-static void peep_update_ride_no_free_vehicle_rejoin_queue(Guest* peep, Ride& ride)
-{
-    TileCoordsXYZD entranceLocation = ride.GetStation(peep->CurrentRideStation).Entrance;
-
-    int32_t x = entranceLocation.x * 32;
-    int32_t y = entranceLocation.y * 32;
-    x += 16 - DirectionOffsets[entranceLocation.direction].x * 20;
-    y += 16 - DirectionOffsets[entranceLocation.direction].y * 20;
-
-    peep->SetDestination({ x, y }, 2);
-    peep->SetState(PeepState::QueuingFront);
-    peep->RideSubState = PeepRideSubState::AtEntrance;
-
-    ride.QueueInsertGuestAtFront(peep->CurrentRideStation, peep);
-}
-
-/**
- *
- *  rct2: 0x00691E42
- * Note: Before this was the entry
- * point for sub state 1 and 3. The
- * check has been removed that would
- * branch it out to 1 and 3. Now uses
- * separate functions.
- */
-void Guest::UpdateRideFreeVehicleCheck()
-{
-    auto ride = get_ride(CurrentRide);
-    if (ride == nullptr)
-        return;
-
-    if (ride->GetRideTypeDescriptor().HasFlag(RIDE_TYPE_FLAG_NO_VEHICLES))
-    {
-        if (ride->status != RideStatus::Open || ride->vehicle_change_timeout != 0 || (++RejoinQueueTimeout) == 0)
-        {
-            peep_update_ride_no_free_vehicle_rejoin_queue(this, *ride);
-            return;
-        }
-
-        UpdateRideFreeVehicleEnterRide(*ride);
-        return;
-    }
-
-    Vehicle* vehicle = GetEntity<Vehicle>(ride->vehicles[CurrentTrain]);
-    if (vehicle == nullptr)
-    {
-        // TODO: Leave ride on failure goes for all returns on nullptr in this function
-        return;
-    }
-    vehicle = vehicle->GetCar(CurrentCar);
-    if (vehicle == nullptr)
-        return;
-
-    rct_ride_entry* ride_entry = vehicle->GetRideEntry();
-    if (ride_entry == nullptr)
-    {
-        return;
-    }
-
-    if (ride_entry->Cars[0].flags & CAR_ENTRY_FLAG_MINI_GOLF)
-    {
-        vehicle->mini_golf_flags &= ~MiniGolfFlag::Flag5;
-
-        for (size_t i = 0; i < ride->NumTrains; ++i)
-        {
-            Vehicle* train = GetEntity<Vehicle>(ride->vehicles[i]);
-            if (train == nullptr)
-                continue;
-
-            Vehicle* second_vehicle = GetEntity<Vehicle>(train->next_vehicle_on_train);
-            if (second_vehicle == nullptr)
-                continue;
-
-            if (second_vehicle->num_peeps == 0)
-                continue;
-
-            if (second_vehicle->mini_golf_flags & MiniGolfFlag::Flag5)
-                continue;
-
-            return;
-        }
-    }
-
-    if (!vehicle->IsUsedInPairs())
-    {
-        UpdateRideFreeVehicleEnterRide(*ride);
-        return;
-    }
-
-    if (ride->mode == RideMode::ForwardRotation || ride->mode == RideMode::BackwardRotation)
-    {
-        if (CurrentSeat & 1 || !(vehicle->next_free_seat & 1))
-        {
-            UpdateRideFreeVehicleEnterRide(*ride);
-            return;
-        }
-    }
-    else
-    {
-        uint8_t seat = CurrentSeat | 1;
-        if (seat < vehicle->next_free_seat)
-        {
-            UpdateRideFreeVehicleEnterRide(*ride);
-            return;
-        }
-    }
-
-    Vehicle* currentTrain = GetEntity<Vehicle>(ride->vehicles[CurrentTrain]);
-    if (currentTrain == nullptr)
-    {
-        return;
-    }
-    if (ride->status == RideStatus::Open && ++RejoinQueueTimeout != 0
-        && !currentTrain->HasUpdateFlag(VEHICLE_UPDATE_FLAG_TRAIN_READY_DEPART))
-    {
-        return;
-    }
-
-    if (ride->mode != RideMode::ForwardRotation && ride->mode != RideMode::BackwardRotation)
-    {
-        if (vehicle->next_free_seat - 1 != CurrentSeat)
-            return;
-    }
-
-    vehicle->next_free_seat--;
-    vehicle->peep[CurrentSeat] = EntityId::GetNull();
-
-    peep_update_ride_no_free_vehicle_rejoin_queue(this, *ride);
-}
-
-void Guest::UpdateRideApproachVehicle()
-{
-    if (auto loc = UpdateAction(); loc.has_value())
-    {
-        MoveTo({ loc.value(), z });
-        return;
-    }
-    RideSubState = PeepRideSubState::EnterVehicle;
-}
-
-void Guest::UpdateRideEnterVehicle()
-{
-    auto* ride = get_ride(CurrentRide);
-    if (ride != nullptr)
-    {
-        auto* vehicle = GetEntity<Vehicle>(ride->vehicles[CurrentTrain]);
-        if (vehicle != nullptr)
-        {
-            vehicle = vehicle->GetCar(CurrentCar);
-            if (vehicle == nullptr)
-            {
-                return;
-            }
-
-            if (ride->mode != RideMode::ForwardRotation && ride->mode != RideMode::BackwardRotation)
-            {
-                if (CurrentSeat != vehicle->num_peeps)
-                    return;
-            }
-
-            if (vehicle->IsUsedInPairs())
-            {
-                auto* seatedGuest = GetEntity<Guest>(vehicle->peep[CurrentSeat ^ 1]);
-                if (seatedGuest != nullptr)
-                {
-                    if (seatedGuest->RideSubState != PeepRideSubState::EnterVehicle)
-                        return;
-
-                    vehicle->num_peeps++;
-                    ride->cur_num_customers++;
-
-                    vehicle->ApplyMass(seatedGuest->Mass);
-                    seatedGuest->MoveTo({ LOCATION_NULL, 0, 0 });
-                    seatedGuest->SetState(PeepState::OnRide);
-                    seatedGuest->GuestTimeOnRide = 0;
-                    seatedGuest->RideSubState = PeepRideSubState::OnRide;
-                    seatedGuest->OnEnterRide(*ride);
-                }
-            }
-
-            vehicle->num_peeps++;
-            ride->cur_num_customers++;
-
-            vehicle->ApplyMass(Mass);
-            vehicle->Invalidate();
-
-            MoveTo({ LOCATION_NULL, 0, 0 });
-
-            SetState(PeepState::OnRide);
-
-            GuestTimeOnRide = 0;
-            RideSubState = PeepRideSubState::OnRide;
-            OnEnterRide(*ride);
-        }
-    }
-}
-
-/**
- *
- *  rct2: 0x00693028
- */
-void Guest::UpdateRideLeaveVehicle()
-{
-    auto ride = get_ride(CurrentRide);
-    if (ride == nullptr)
-        return;
-
-    Vehicle* vehicle = GetEntity<Vehicle>(ride->vehicles[CurrentTrain]);
-    if (vehicle == nullptr)
-        return;
-
-    StationIndex ride_station = vehicle->current_station;
-    vehicle = vehicle->GetCar(CurrentCar);
-    if (vehicle == nullptr)
-    {
-        return;
-    }
-
-    // Check if ride is NOT Ferris Wheel.
-    if (ride->mode != RideMode::ForwardRotation && ride->mode != RideMode::BackwardRotation)
-    {
-        if (vehicle->num_peeps - 1 != CurrentSeat)
-            return;
-    }
-
-    ActionSpriteImageOffset++;
-    if (ActionSpriteImageOffset & 3)
-        return;
-
-    ActionSpriteImageOffset = 0;
-
-    vehicle->num_peeps--;
-    vehicle->ApplyMass(-Mass);
-    vehicle->Invalidate();
-
-    if (ride_station.ToUnderlying() >= OpenRCT2::Limits::MaxStationsPerRide)
-    {
-        // HACK #5658: Some parks have hacked rides which end up in this state
-        auto bestStationIndex = ride_get_first_valid_station_exit(*ride);
-        if (bestStationIndex.IsNull())
-        {
-            bestStationIndex = StationIndex::FromUnderlying(0);
-        }
-        ride_station = bestStationIndex;
-    }
-    CurrentRideStation = ride_station;
-    rct_ride_entry* rideEntry = vehicle->GetRideEntry();
-    if (rideEntry == nullptr)
-    {
-        return;
-    }
-
-    CarEntry* carEntry = &rideEntry->Cars[vehicle->vehicle_type];
-
-    assert(CurrentRideStation.ToUnderlying() < OpenRCT2::Limits::MaxStationsPerRide);
-    auto& station = ride->GetStation(CurrentRideStation);
-
-    if (!(carEntry->flags & CAR_ENTRY_FLAG_LOADING_WAYPOINTS))
-    {
-        TileCoordsXYZD exitLocation = station.Exit;
-        CoordsXYZD platformLocation;
-        platformLocation.z = station.GetBaseZ();
-
-        platformLocation.direction = DirectionReverse(exitLocation.direction);
-
-        if (!ride->GetRideTypeDescriptor().HasFlag(RIDE_TYPE_FLAG_VEHICLE_IS_INTEGRAL))
-        {
-            for (; vehicle != nullptr && !vehicle->IsHead(); vehicle = GetEntity<Vehicle>(vehicle->prev_vehicle_on_ride))
-            {
-                auto trackType = vehicle->GetTrackType();
-                if (trackType == TrackElemType::Flat || trackType > TrackElemType::MiddleStation)
-                    continue;
-
-                bool foundStation = false;
-                for (auto* trackElement : TileElementsView<TrackElement>(vehicle->TrackLocation))
-                {
-                    if (trackElement->GetBaseZ() != vehicle->TrackLocation.z)
-                        continue;
-
-                    if (trackElement->GetStationIndex() != CurrentRideStation)
-                        continue;
-
-                    foundStation = true;
-                    break;
-                }
-
-                if (foundStation)
-                    break;
-            }
-
-            if (vehicle == nullptr)
-            {
-                return;
-            }
-            uint8_t shiftMultiplier = 12;
-            uint8_t specialDirection = platformLocation.direction;
-
-            rideEntry = get_ride_entry(ride->subtype);
-
-            if (rideEntry != nullptr)
-            {
-                carEntry = &rideEntry->Cars[rideEntry->DefaultCar];
-
-                if (carEntry->flags & CAR_ENTRY_FLAG_GO_KART)
-                {
-                    shiftMultiplier = 9;
-                }
-
-                if (carEntry->flags & (CAR_ENTRY_FLAG_CHAIRLIFT | CAR_ENTRY_FLAG_GO_KART))
-                {
-                    specialDirection = ((vehicle->sprite_direction + 3) / 8) + 1;
-                    specialDirection &= 3;
-
-                    if (vehicle->TrackSubposition == VehicleTrackSubposition::GoKartsRightLane)
-                        specialDirection = DirectionReverse(specialDirection);
-                }
-            }
-
-            int16_t xShift = DirectionOffsets[specialDirection].x;
-            int16_t yShift = DirectionOffsets[specialDirection].y;
-
-            platformLocation.x = vehicle->x + xShift * shiftMultiplier;
-            platformLocation.y = vehicle->y + yShift * shiftMultiplier;
-
-            peep_go_to_ride_exit(
-                this, *ride, platformLocation.x, platformLocation.y, platformLocation.z, platformLocation.direction);
-            return;
-        }
-
-        platformLocation.x = vehicle->x + DirectionOffsets[platformLocation.direction].x * 12;
-        platformLocation.y = vehicle->y + DirectionOffsets[platformLocation.direction].y * 12;
-
-        // This can evaluate to false with buggy custom rides.
-        if (CurrentSeat < carEntry->peep_loading_positions.size())
-        {
-            int8_t loadPosition = carEntry->peep_loading_positions[CurrentSeat];
-
-            switch (vehicle->sprite_direction / 8)
-            {
-                case 0:
-                    platformLocation.x -= loadPosition;
-                    break;
-                case 1:
-                    platformLocation.y += loadPosition;
-                    break;
-                case 2:
-                    platformLocation.x += loadPosition;
-                    break;
-                case 3:
-                    platformLocation.y -= loadPosition;
-                    break;
-            }
-        }
-        else
-        {
-            log_verbose(
-                "CurrentSeat %d is too large! (Vehicle entry has room for %d.)", CurrentSeat,
-                carEntry->peep_loading_positions.size());
-        }
-
-        platformLocation.z = station.GetBaseZ();
-
-        peep_go_to_ride_exit(
-            this, *ride, platformLocation.x, platformLocation.y, platformLocation.z, platformLocation.direction);
-        return;
-    }
-
-    auto exitLocation = station.Exit.ToCoordsXYZD();
-    Guard::Assert(!exitLocation.IsNull());
-
-    TileElement* trackElement = ride_get_station_start_track_element(*ride, CurrentRideStation);
-
-    Direction station_direction = (trackElement == nullptr ? 0 : trackElement->GetDirection());
-
-    vehicle = GetEntity<Vehicle>(ride->vehicles[CurrentTrain]);
-    if (vehicle == nullptr)
-    {
-        return;
-    }
-
-    CoordsXYZ waypointLoc;
-    const auto& rtd = ride->GetRideTypeDescriptor();
-    waypointLoc = { rtd.GetGuestWaypointLocation(*vehicle, *ride, CurrentRideStation),
-                    exitLocation.z + ride->GetRideTypeDescriptor().Heights.PlatformHeight };
-
-    rideEntry = vehicle->GetRideEntry();
-    carEntry = &rideEntry->Cars[vehicle->vehicle_type];
-    if (carEntry == nullptr)
-        return;
-
-    Var37 = ((exitLocation.direction | GetWaypointedSeatLocation(*ride, carEntry, station_direction) * 4) * 4) | 1;
-    Guard::Assert(carEntry->peep_loading_waypoints.size() >= static_cast<size_t>(Var37 / 4));
-    CoordsXYZ exitWaypointLoc = waypointLoc;
-
-    exitWaypointLoc.x += carEntry->peep_loading_waypoints[Var37 / 4][2].x;
-    exitWaypointLoc.y += carEntry->peep_loading_waypoints[Var37 / 4][2].y;
-
-    if (ride->type == RIDE_TYPE_MOTION_SIMULATOR)
-        exitWaypointLoc.z += 15;
-
-    MoveTo(exitWaypointLoc);
-
-    waypointLoc.x += carEntry->peep_loading_waypoints[Var37 / 4][1].x;
-    waypointLoc.y += carEntry->peep_loading_waypoints[Var37 / 4][1].y;
-
-    SetDestination(waypointLoc, 2);
-    RideSubState = PeepRideSubState::ApproachExitWaypoints;
-}
-
-/**
- *
- *  rct2: 0x0069376A
- */
-void Guest::UpdateRidePrepareForExit()
-{
-    auto ride = get_ride(CurrentRide);
-    if (ride == nullptr || CurrentRideStation.ToUnderlying() >= std::size(ride->GetStations()))
-        return;
-
-    auto exit = ride->GetStation(CurrentRideStation).Exit;
-    auto newDestination = exit.ToCoordsXY().ToTileCentre();
-    auto xShift = DirectionOffsets[exit.direction].x;
-    auto yShift = DirectionOffsets[exit.direction].y;
-
-    int16_t shiftMultiplier = 20;
-
-    rct_ride_entry* rideEntry = ride->GetRideEntry();
-    if (rideEntry != nullptr)
-    {
-        CarEntry* carEntry = &rideEntry->Cars[rideEntry->DefaultCar];
-        if (carEntry->flags & (CAR_ENTRY_FLAG_CHAIRLIFT | CAR_ENTRY_FLAG_GO_KART))
-        {
-            shiftMultiplier = 32;
-        }
-    }
-
-    xShift *= shiftMultiplier;
-    yShift *= shiftMultiplier;
-
-    newDestination.x -= xShift;
-    newDestination.y -= yShift;
-
-    SetDestination(newDestination, 2);
-    RideSubState = PeepRideSubState::InExit;
-}
-
-/**
- *
- *  rct2: 0x0069374F
- */
-void Guest::UpdateRideApproachExit()
-{
-    if (auto loc = UpdateAction(); loc.has_value())
-    {
-        MoveTo({ loc.value(), z });
-        return;
-    }
-
-    UpdateRidePrepareForExit();
-}
-
-/**
- *
- *  rct2: 0x0069382E
- */
-void Guest::UpdateRideInExit()
-{
-    auto ride = get_ride(CurrentRide);
-    if (ride == nullptr)
-        return;
-
-    int16_t xy_distance;
-
-    if (auto loc = UpdateAction(xy_distance); loc.has_value())
-    {
-        if (xy_distance >= 16)
-        {
-            int16_t actionZ = ride->GetStation(CurrentRideStation).GetBaseZ();
-
-            actionZ += ride->GetRideTypeDescriptor().Heights.PlatformHeight;
-            MoveTo({ loc.value(), actionZ });
-            return;
-        }
-
-        SwitchToSpecialSprite(0);
-        MoveTo({ loc.value(), z });
-    }
-
-    if (ride->lifecycle_flags & RIDE_LIFECYCLE_ON_RIDE_PHOTO)
-    {
-        ShopItem secondaryItem = ride->GetRideTypeDescriptor().PhotoItem;
-        if (DecideAndBuyItem(*ride, secondaryItem, ride->price[1]))
-        {
-            ride->no_secondary_items_sold++;
-        }
-    }
-    RideSubState = PeepRideSubState::LeaveExit;
-}
-#pragma warning(default : 6011)
-
-CoordsXY GetGuestWaypointLocationDefault(const Vehicle& vehicle, const Ride& ride, const StationIndex& CurrentRideStation)
-{
-    return ride.GetStation(CurrentRideStation).Start.ToTileCentre();
-}
-
-CoordsXY GetGuestWaypointLocationEnterprise(const Vehicle& vehicle, const Ride& ride, const StationIndex& CurrentRideStation)
-{
-    return { vehicle.x, vehicle.y };
-}
-
-/**
- *
- *  rct2: 0x006926AD
- */
-void Guest::UpdateRideApproachVehicleWaypoints()
-{
-    auto ride = get_ride(CurrentRide);
-    if (ride == nullptr)
-        return;
-
-    int16_t xy_distance;
-    uint8_t waypoint = Var37 & 3;
-
-    const auto& rtd = ride->GetRideTypeDescriptor();
-    if (auto loc = UpdateAction(xy_distance); loc.has_value())
-    {
-        rtd.UpdateRideApproachVehicleWaypoints(*this, loc.value(), xy_distance);
-        return;
-    }
-
-    if (waypoint == 2)
-    {
-        RideSubState = PeepRideSubState::EnterVehicle;
-        return;
-    }
-
-    waypoint++;
-    // This is incrementing the actual peep waypoint
-    Var37++;
-
-    Vehicle* vehicle = GetEntity<Vehicle>(ride->vehicles[CurrentTrain]);
-    if (vehicle == nullptr)
-    {
-        return;
-    }
-
-    CoordsXY targetLoc = rtd.GetGuestWaypointLocation(*vehicle, *ride, CurrentRideStation);
-
-    rct_ride_entry* ride_entry = vehicle->GetRideEntry();
-    if (ride_entry == nullptr)
-    {
-        return;
-    }
-
-    CarEntry* vehicle_type = &ride_entry->Cars[vehicle->vehicle_type];
-    Guard::Assert(waypoint < 3);
-    targetLoc.x += vehicle_type->peep_loading_waypoints[Var37 / 4][waypoint].x;
-    targetLoc.y += vehicle_type->peep_loading_waypoints[Var37 / 4][waypoint].y;
-
-    SetDestination(targetLoc);
-}
-
-void UpdateRideApproachVehicleWaypointsMotionSimulator(Guest& guest, const CoordsXY& loc, int16_t& xy_distance)
-{
-    int16_t actionZ;
-    auto ride = get_ride(guest.CurrentRide);
-    // Motion simulators have steps this moves the peeps up the steps
-    actionZ = ride->GetStation(guest.CurrentRideStation).GetBaseZ() + 2;
-
-    uint8_t waypoint = guest.Var37 & 3;
-    if (waypoint == 2)
-    {
-        xy_distance -= 12;
-        if (xy_distance < 0)
-            xy_distance = 0;
-
-        if (xy_distance <= 15)
-        {
-            actionZ += 15 - xy_distance;
-        }
-    }
-    guest.MoveTo({ loc, actionZ });
-}
-
-void UpdateRideApproachVehicleWaypointsDefault(Guest& guest, const CoordsXY& loc, int16_t& xy_distance)
-{
-    guest.MoveTo({ loc, guest.z });
-}
-
-/**
- *
- *  rct2: 0x0069357D
- */
-void Guest::UpdateRideApproachExitWaypoints()
-{
-    auto ride = get_ride(CurrentRide);
-    if (ride == nullptr)
-        return;
-
-    int16_t xy_distance;
-
-    if (auto loc = UpdateAction(xy_distance); loc.has_value())
-    {
-        int16_t actionZ;
-
-        if (ride->type == RIDE_TYPE_MOTION_SIMULATOR)
-        {
-            actionZ = ride->GetStation(CurrentRideStation).GetBaseZ() + 2;
-
-            if ((Var37 & 3) == 1)
-            {
-                if (xy_distance > 15)
-                    xy_distance = 15;
-
-                actionZ += xy_distance;
-            }
-        }
-        else
-        {
-            actionZ = z;
-        }
-        MoveTo({ loc.value(), actionZ });
-        return;
-    }
-
-    if ((Var37 & 3) != 0)
-    {
-        if ((Var37 & 3) == 3)
-        {
-            UpdateRidePrepareForExit();
-            return;
-        }
-
-        Var37--;
-        Vehicle* vehicle = GetEntity<Vehicle>(ride->vehicles[CurrentTrain]);
-        if (vehicle == nullptr)
-        {
-            return;
-        }
-
-        const auto& rtd = ride->GetRideTypeDescriptor();
-        CoordsXY targetLoc = rtd.GetGuestWaypointLocation(*vehicle, *ride, CurrentRideStation);
-
-        rct_ride_entry* rideEntry = vehicle->GetRideEntry();
-        CarEntry* carEntry = &rideEntry->Cars[vehicle->vehicle_type];
-
-        Guard::Assert((Var37 & 3) < 3);
-        targetLoc.x += carEntry->peep_loading_waypoints[Var37 / 4][Var37 & 3].x;
-        targetLoc.y += carEntry->peep_loading_waypoints[Var37 / 4][Var37 & 3].y;
-
-        SetDestination(targetLoc);
-        return;
-    }
-
-    Var37 |= 3;
-
-    auto targetLoc = ride->GetStation(CurrentRideStation).Exit.ToCoordsXYZD().ToTileCentre();
-    uint8_t exit_direction = DirectionReverse(targetLoc.direction);
-
-    int16_t x_shift = DirectionOffsets[exit_direction].x;
-    int16_t y_shift = DirectionOffsets[exit_direction].y;
-
-    int16_t shift_multiplier = 20;
-
-    auto rideEntry = get_ride_entry(ride->subtype);
-    if (rideEntry != nullptr)
-    {
-        auto carEntry = &rideEntry->Cars[rideEntry->DefaultCar];
-        if (carEntry->flags & (CAR_ENTRY_FLAG_CHAIRLIFT | CAR_ENTRY_FLAG_GO_KART))
-        {
-            shift_multiplier = 32;
-        }
-    }
-
-    x_shift *= shift_multiplier;
-    y_shift *= shift_multiplier;
-
-    targetLoc.x -= x_shift;
-    targetLoc.y -= y_shift;
-
-    SetDestination(targetLoc);
-}
-
-/**
- *
- *  rct2: 0x006927B3
- */
-void Guest::UpdateRideApproachSpiralSlide()
-{
-    auto ride = get_ride(CurrentRide);
-    if (ride == nullptr)
-        return;
-
-    if (auto loc = UpdateAction(); loc.has_value())
-    {
-        MoveTo({ loc.value(), z });
-        return;
-    }
-
-    uint8_t waypoint = Var37 & 3;
-
-    if (waypoint == 3)
-    {
-        SubState = 15;
-        SetDestination({ 0, 0 });
-        Var37 = (Var37 / 4) & 0xC;
-        MoveTo({ LOCATION_NULL, y, z });
-        return;
-    }
-
-    [[maybe_unused]] const auto& rtd = ride->GetRideTypeDescriptor();
-    if (waypoint == 2)
-    {
-        bool lastRide = false;
-        if (ride->status != RideStatus::Open)
-            lastRide = true;
-        else if (CurrentCar++ != 0)
-        {
-            if (ride->mode == RideMode::SingleRidePerAdmission)
-                lastRide = true;
-            if (static_cast<uint8_t>(CurrentCar - 1) > (scenario_rand() & 0xF))
-                lastRide = true;
-        }
-
-        if (lastRide)
-        {
-            auto exit = ride->GetStation(CurrentRideStation).Exit;
-            waypoint = 1;
-            Var37 = (exit.direction * 4) | (Var37 & 0x30) | waypoint;
-            CoordsXY targetLoc = ride->GetStation(CurrentRideStation).Start;
-
-            assert(rtd.HasFlag(RIDE_TYPE_FLAG_IS_SPIRAL_SLIDE));
-            targetLoc += SpiralSlideWalkingPath[Var37];
-
-            SetDestination(targetLoc);
-            RideSubState = PeepRideSubState::LeaveSpiralSlide;
-            return;
-        }
-    }
-    waypoint++;
-    // Actually increment the real peep waypoint
-    Var37++;
-
-    CoordsXY targetLoc = ride->GetStation(CurrentRideStation).Start;
-
-    assert(rtd.HasFlag(RIDE_TYPE_FLAG_IS_SPIRAL_SLIDE));
-    targetLoc += SpiralSlideWalkingPath[Var37];
-
-    SetDestination(targetLoc);
-}
-
-/** rct2: 0x00981F0C, 0x00981F0E */
-static constexpr const CoordsXY _SpiralSlideEnd[] = {
-    { 25, 56 },
-    { 56, 7 },
-    { 7, -24 },
-    { -24, 25 },
-};
-
-/** rct2: 0x00981F1C, 0x00981F1E */
-static constexpr const CoordsXY _SpiralSlideEndWaypoint[] = {
-    { 8, 56 },
-    { 56, 24 },
-    { 24, -24 },
-    { -24, 8 },
-};
-
-/**
- *
- *  rct2: 0x00692D83
- */
-void Guest::UpdateRideOnSpiralSlide()
-{
-    auto ride = get_ride(CurrentRide);
-
-    if (ride == nullptr)
-        return;
-
-    const auto& rtd = ride->GetRideTypeDescriptor();
-    if (!rtd.HasFlag(RIDE_TYPE_FLAG_IS_SPIRAL_SLIDE))
-        return;
-
-    auto destination = GetDestination();
-    if ((Var37 & 3) == 0)
-    {
-        switch (destination.x)
-        {
-            case 0:
-                destination.y++;
-                if (destination.y >= 30)
-                    destination.x++;
-
-                SetDestination(destination);
-                return;
-            case 1:
-                if (ride->slide_in_use != 0)
-                    return;
-
-                ride->slide_in_use++;
-                ride->slide_peep = sprite_index;
-                ride->slide_peep_t_shirt_colour = TshirtColour;
-                ride->spiral_slide_progress = 0;
-                destination.x++;
-
-                SetDestination(destination);
-                return;
-            case 2:
-                return;
-            case 3:
-            {
-                auto newLocation = ride->GetStation(CurrentRideStation).Start;
-                uint8_t dir = (Var37 / 4) & 3;
-
-                // Set the location that the peep walks to go on slide again
-                destination = newLocation + _SpiralSlideEndWaypoint[dir];
-                SetDestination(destination);
-
-                // Move the peep sprite to just at the end of the slide
-                newLocation.x += _SpiralSlideEnd[dir].x;
-                newLocation.y += _SpiralSlideEnd[dir].y;
-
-                MoveTo({ newLocation, z });
-
-                sprite_direction = (Var37 & 0xC) * 2;
-
-                Var37++;
-                return;
-            }
-            default:
-                return;
-        }
-    }
-
-    if (auto loc = UpdateAction(); loc.has_value())
-    {
-        MoveTo({ loc.value(), z });
-        return;
-    }
-
-    uint8_t waypoint = 2;
-    Var37 = (Var37 * 4 & 0x30) + waypoint;
-
-    CoordsXY targetLoc = ride->GetStation(CurrentRideStation).Start;
-
-    targetLoc += SpiralSlideWalkingPath[Var37];
-
-    SetDestination(targetLoc);
-    RideSubState = PeepRideSubState::ApproachSpiralSlide;
-}
-
-/**
- *
- *  rct2: 0x00692C6B
- */
-void Guest::UpdateRideLeaveSpiralSlide()
-{
-    // Iterates through the spiral slide waypoints until it reaches
-    // waypoint 0. Then it readies to leave the ride by the entrance.
-    if (auto loc = UpdateAction(); loc.has_value())
-    {
-        MoveTo({ loc.value(), z });
-        return;
-    }
-
-    auto ride = get_ride(CurrentRide);
-    if (ride == nullptr)
-        return;
-
-    uint8_t waypoint = Var37 & 3;
-
-    if (waypoint != 0)
-    {
-        if (waypoint == 3)
-        {
-            UpdateRidePrepareForExit();
-            return;
-        }
-
-        waypoint--;
-        // Actually decrement the peep waypoint
-        Var37--;
-        CoordsXY targetLoc = ride->GetStation(CurrentRideStation).Start;
-
-        [[maybe_unused]] const auto& rtd = ride->GetRideTypeDescriptor();
-        assert(rtd.HasFlag(RIDE_TYPE_FLAG_IS_SPIRAL_SLIDE));
-        targetLoc += SpiralSlideWalkingPath[Var37];
-
-        SetDestination(targetLoc);
-        return;
-    }
-
-    // Actually force the final waypoint
-    Var37 |= 3;
-
-    auto targetLoc = ride->GetStation(CurrentRideStation).Exit.ToCoordsXYZD().ToTileCentre();
-
-    int16_t xShift = DirectionOffsets[DirectionReverse(targetLoc.direction)].x;
-    int16_t yShift = DirectionOffsets[DirectionReverse(targetLoc.direction)].y;
-
-    int16_t shiftMultiplier = 20;
-
-    xShift *= shiftMultiplier;
-    yShift *= shiftMultiplier;
-
-    targetLoc.x -= xShift;
-    targetLoc.y -= yShift;
-
-    SetDestination(targetLoc);
-}
-
-/** rct2: 0x00981FE4 */
-static constexpr const uint8_t _MazeGetNewDirectionFromEdge[][4] = {
-    { 15, 7, 15, 7 },
-    { 11, 3, 11, 3 },
-    { 7, 15, 7, 15 },
-    { 3, 11, 3, 11 },
-};
-
-/** rct2: 0x00981FF4 */
-static constexpr const uint8_t _MazeCurrentDirectionToOpenHedge[][4] = {
-    { 1, 2, 14, 0 },
-    { 4, 5, 6, 2 },
-    { 6, 8, 9, 10 },
-    { 14, 10, 12, 13 },
-};
-
-/**
- *
- *  rct2: 0x00692A83
- */
-void Guest::UpdateRideMazePathfinding()
-{
-    if (auto loc = UpdateAction(); loc.has_value())
-    {
-        MoveTo({ loc.value(), z });
-        return;
-    }
-
-    auto ride = get_ride(CurrentRide);
-    if (ride == nullptr)
-        return;
-
-    if (Var37 == 16)
-    {
-        UpdateRidePrepareForExit();
-        return;
-    }
-
-    if (IsActionInterruptable())
-    {
-        if (Energy > 80 && !(PeepFlags & PEEP_FLAGS_SLOW_WALK) && !ClimateIsRaining() && (scenario_rand() & 0xFFFF) <= 2427)
-        {
-            Action = PeepActionType::Jump;
-            ActionFrame = 0;
-            ActionSpriteImageOffset = 0;
-            UpdateCurrentActionSpriteType();
-        }
-    }
-
-    auto targetLoc = GetDestination().ToTileStart();
-
-    auto stationBaseZ = ride->GetStation().GetBaseZ();
-
-    // Find the station track element
-    auto trackElement = MapGetTrackElementAt({ targetLoc, stationBaseZ });
-    if (trackElement == nullptr)
-    {
-        return;
-    }
-
-    uint16_t mazeEntry = trackElement->GetMazeEntry();
-    // Var37 is 3, 7, 11 or 15
-    uint8_t hedges[4]{ 0xFF, 0xFF, 0xFF, 0xFF };
-    uint8_t openCount = 0;
-    uint8_t mazeReverseLastEdge = DirectionReverse(MazeLastEdge);
-    for (uint8_t i = 0; i < NumOrthogonalDirections; ++i)
-    {
-        if (!(mazeEntry & (1 << _MazeCurrentDirectionToOpenHedge[Var37 / 4][i])) && i != mazeReverseLastEdge)
-        {
-            hedges[openCount++] = i;
-        }
-    }
-
-    if (openCount == 0)
-    {
-        if ((mazeEntry & (1 << _MazeCurrentDirectionToOpenHedge[Var37 / 4][mazeReverseLastEdge])))
-        {
-            return;
-        }
-        hedges[openCount++] = mazeReverseLastEdge;
-    }
-
-    uint8_t chosenEdge = hedges[scenario_rand() % openCount];
-    assert(chosenEdge != 0xFF);
-
-    targetLoc = GetDestination() + CoordsDirectionDelta[chosenEdge] / 2;
-
-    enum class maze_type
-    {
-        invalid,
-        hedge,
-        entrance_or_exit
-    };
-    maze_type mazeType = maze_type::invalid;
-
-    auto tileElement = MapGetFirstElementAt(targetLoc);
-    if (tileElement == nullptr)
-        return;
-    do
-    {
-        if (stationBaseZ != tileElement->GetBaseZ())
-            continue;
-
-        if (tileElement->GetType() == TileElementType::Track)
-        {
-            mazeType = maze_type::hedge;
-            break;
-        }
-
-        if (tileElement->GetType() == TileElementType::Entrance
-            && tileElement->AsEntrance()->GetEntranceType() == ENTRANCE_TYPE_RIDE_EXIT)
-        {
-            mazeType = maze_type::entrance_or_exit;
-            break;
-        }
-    } while (!(tileElement++)->IsLastForTile());
-
-    switch (mazeType)
-    {
-        case maze_type::invalid:
-            MazeLastEdge++;
-            MazeLastEdge &= 3;
-            return;
-        case maze_type::hedge:
-            SetDestination(targetLoc);
-            Var37 = _MazeGetNewDirectionFromEdge[Var37 / 4][chosenEdge];
-            MazeLastEdge = chosenEdge;
-            break;
-        case maze_type::entrance_or_exit:
-            targetLoc = GetDestination();
-            if (chosenEdge & 1)
-            {
-                targetLoc.x = targetLoc.ToTileCentre().x;
-            }
-            else
-            {
-                targetLoc.y = targetLoc.ToTileCentre().y;
-            }
-            SetDestination(targetLoc);
-            Var37 = 16;
-            MazeLastEdge = chosenEdge;
-            break;
-    }
-
-    if (auto loc = UpdateAction(); loc.has_value())
-    {
-        MoveTo({ loc.value(), z });
-        return;
-    }
-}
-
-/**
- *
- *  rct2: 0x006938D2
- */
-void Guest::UpdateRideLeaveExit()
-{
-    auto ride = get_ride(CurrentRide);
-
-    if (auto loc = UpdateAction(); loc.has_value())
-    {
-        if (ride != nullptr)
-        {
-            MoveTo({ loc.value(), ride->GetStation(CurrentRideStation).GetBaseZ() });
-        }
-        return;
-    }
-
-    OnExitRide(*ride);
-
-    if (ride != nullptr && (PeepFlags & PEEP_FLAGS_TRACKING))
-    {
-        auto ft = Formatter();
-        FormatNameTo(ft);
-        ride->FormatNameTo(ft);
-
-        if (gConfigNotifications.GuestLeftRide)
-        {
-            News::AddItemToQueue(News::ItemType::PeepOnRide, STR_PEEP_TRACKING_LEFT_RIDE_X, sprite_index, ft);
-        }
-    }
-
-    InteractionRideIndex = RideId::GetNull();
-    SetState(PeepState::Falling);
-
-    CoordsXY targetLoc = { x, y };
-
-    // Find the station track element
-    for (auto* pathElement : TileElementsView<PathElement>(targetLoc))
-    {
-        int16_t height = MapHeightFromSlope(targetLoc, pathElement->GetSlopeDirection(), pathElement->IsSloped());
-        height += pathElement->GetBaseZ();
-
-        int16_t z_diff = z - height;
-        if (z_diff > 0 || z_diff < -16)
-            continue;
-
-        MoveTo({ x, y, height });
-        return;
-    }
-}
-
-/**
- *
- *  rct2: 0x0069299C
- */
-void Guest::UpdateRideShopApproach()
-{
-    if (auto loc = UpdateAction(); loc.has_value())
-    {
-        MoveTo({ loc.value(), z });
-        return;
-    }
-
-    RideSubState = PeepRideSubState::InteractShop;
-}
-
-/**
- *
- *  rct2: 0x006929BB
- */
-void Guest::UpdateRideShopInteract()
-{
-    auto ride = get_ride(CurrentRide);
-    if (ride == nullptr)
-        return;
-
-    const int16_t tileCentreX = NextLoc.x + 16;
-    const int16_t tileCentreY = NextLoc.y + 16;
-
-    const auto& rtd = ride->GetRideTypeDescriptor();
-    if (rtd.HasFlag(RIDE_TYPE_FLAG_IS_FIRST_AID))
-    {
-        if (Nausea <= 35)
-        {
-            RideSubState = PeepRideSubState::LeaveShop;
-
-            SetDestination({ tileCentreX, tileCentreY }, 3);
-            HappinessTarget = std::min(HappinessTarget + 30, PEEP_MAX_HAPPINESS);
-            Happiness = HappinessTarget;
-        }
-        else
-        {
-            Nausea--;
-            NauseaTarget = Nausea;
-        }
-        return;
-    }
-
-    if (Toilet != 0)
-    {
-        Toilet--;
-        return;
-    }
-
-    // Do not play toilet flush sound on title screen as it's considered loud and annoying
-    if (!(gScreenFlags & SCREEN_FLAGS_TITLE_DEMO))
-    {
-        OpenRCT2::Audio::Play3D(OpenRCT2::Audio::SoundId::ToiletFlush, GetLocation());
-    }
-
-    RideSubState = PeepRideSubState::LeaveShop;
-
-    SetDestination({ tileCentreX, tileCentreY }, 3);
-
-    HappinessTarget = std::min(HappinessTarget + 30, PEEP_MAX_HAPPINESS);
-    Happiness = HappinessTarget;
-    StopPurchaseThought(ride->type);
-}
-
-/**
- *
- *  rct2: 0x00692935
- */
-void Guest::UpdateRideShopLeave()
-{
-    if (auto loc = UpdateAction(); loc.has_value())
-    {
-        const auto curLoc = GetLocation();
-        MoveTo({ loc.value(), curLoc.z });
-
-        const auto newLoc = GetLocation().ToTileStart();
-        if (newLoc.x != NextLoc.x)
-            return;
-        if (newLoc.y != NextLoc.y)
-            return;
-    }
-
-    // #11758 Previously SetState(PeepState::Walking) caused Peeps to double-back to exit point of shop
-    SetState(PeepState::Falling);
-
-    auto ride = get_ride(CurrentRide);
-    if (ride != nullptr)
-    {
-        ride->total_customers++;
-        ride->window_invalidate_flags |= RIDE_INVALIDATE_RIDE_CUSTOMER;
-        ride->UpdateSatisfaction(Happiness / 64);
-    }
-}
-
 void Guest::UpdateGuest()
 {
     switch (State)
     {
         case PeepState::QueuingFront:
-            UpdateRide();
+            GuestRideHelper(*this).UpdateRide();
             break;
         case PeepState::LeavingRide:
-            UpdateRide();
+            GuestRideHelper(*this).UpdateRide();
             break;
         case PeepState::Walking:
             UpdateWalking();
             break;
         case PeepState::Queuing:
-            UpdateQueuing();
+            GuestRideHelper(*this).UpdateQueuing();
             break;
         case PeepState::EnteringRide:
-            UpdateRide();
+            GuestRideHelper(*this).UpdateRide();
             break;
         case PeepState::Sitting:
             UpdateSitting();
@@ -5127,85 +2112,7 @@ void Guest::UpdateGuest()
     }
 }
 
-/**
- *
- *  rct2: 0x691A30
- * Used by entering_ride and queueing_front */
-void Guest::UpdateRide()
-{
-    NextFlags &= ~PEEP_NEXT_FLAG_IS_SLOPED;
-
-    switch (RideSubState)
-    {
-        case PeepRideSubState::AtEntrance:
-            UpdateRideAtEntrance();
-            break;
-        case PeepRideSubState::InEntrance:
-            UpdateRideAdvanceThroughEntrance();
-            break;
-        case PeepRideSubState::FreeVehicleCheck:
-            UpdateRideFreeVehicleCheck();
-            break;
-        case PeepRideSubState::LeaveEntrance:
-            UpdateRideAdvanceThroughEntrance();
-            break;
-        case PeepRideSubState::ApproachVehicle:
-            UpdateRideApproachVehicle();
-            break;
-        case PeepRideSubState::EnterVehicle:
-            UpdateRideEnterVehicle();
-            break;
-        case PeepRideSubState::OnRide:
-            // No action, on ride.
-            break;
-        case PeepRideSubState::LeaveVehicle:
-            UpdateRideLeaveVehicle();
-            break;
-        case PeepRideSubState::ApproachExit:
-            UpdateRideApproachExit();
-            break;
-        case PeepRideSubState::InExit:
-            UpdateRideInExit();
-            break;
-        case PeepRideSubState::ApproachVehicleWaypoints:
-            UpdateRideApproachVehicleWaypoints();
-            break;
-        case PeepRideSubState::ApproachExitWaypoints:
-            UpdateRideApproachExitWaypoints();
-            break;
-        case PeepRideSubState::ApproachSpiralSlide:
-            UpdateRideApproachSpiralSlide();
-            break;
-        case PeepRideSubState::OnSpiralSlide:
-            UpdateRideOnSpiralSlide();
-            break;
-        case PeepRideSubState::LeaveSpiralSlide:
-            UpdateRideLeaveSpiralSlide();
-            break;
-        case PeepRideSubState::MazePathfinding:
-            UpdateRideMazePathfinding();
-            break;
-        case PeepRideSubState::LeaveExit:
-            UpdateRideLeaveExit();
-            break;
-        case PeepRideSubState::ApproachShop:
-            UpdateRideShopApproach();
-            break;
-        case PeepRideSubState::InteractShop:
-            UpdateRideShopInteract();
-            break;
-        case PeepRideSubState::LeaveShop:
-            UpdateRideShopLeave();
-            break;
-        default:
-            // Invalid peep sub-state
-            assert(false);
-            break;
-    }
-}
-
 static void peep_update_walking_break_scenery(Guest* peep);
-static bool peep_find_ride_to_look_at(Peep* peep, uint8_t edge, RideId* rideToView, uint8_t* rideSeatToView);
 
 /**
  *
@@ -5404,7 +2311,7 @@ void Guest::UpdateWalking()
 
     RideId ride_to_view;
     uint8_t ride_seat_to_view;
-    if (IsOnLevelCrossing() || !peep_find_ride_to_look_at(this, chosen_edge, &ride_to_view, &ride_seat_to_view))
+    if (IsOnLevelCrossing() || !GuestRideHelper(*this).FindRideToLookAt(chosen_edge, &ride_to_view, &ride_seat_to_view))
         return;
 
     // Check if there is a peep watching (and if there is place for us)
@@ -5488,126 +2395,6 @@ void Guest::UpdateWaitingAtCrossing()
     }
 
     UpdateCurrentActionSpriteType();
-}
-
-/**
- *
- *  rct2: 0x69185D
- */
-void Guest::UpdateQueuing()
-{
-    if (!CheckForPath())
-    {
-        RemoveFromQueue();
-        return;
-    }
-    auto ride = get_ride(CurrentRide);
-    if (ride == nullptr || ride->status != RideStatus::Open)
-    {
-        RemoveFromQueue();
-        SetState(PeepState::One);
-        return;
-    }
-
-    // If not in the queue then at front of queue
-    if (RideSubState != PeepRideSubState::InQueue)
-    {
-        bool is_front = true;
-        // Fix #4819: Occasionally the peep->GuestNextInQueue is incorrectly set
-        // to prevent this from causing the peeps to enter a loop
-        // first check if the next in queue is actually nearby
-        // if they are not then it's safe to assume that this is
-        // the front of the queue.
-        Peep* nextGuest = GetEntity<Guest>(GuestNextInQueue);
-        if (nextGuest != nullptr)
-        {
-            if (abs(nextGuest->x - x) < 32 && abs(nextGuest->y - y) < 32)
-            {
-                is_front = false;
-            }
-        }
-
-        if (is_front)
-        {
-            // Happens every time peep goes onto ride.
-            DestinationTolerance = 0;
-            SetState(PeepState::QueuingFront);
-            RideSubState = PeepRideSubState::AtEntrance;
-        }
-
-        return;
-    }
-
-    uint8_t pathingResult;
-    PerformNextAction(pathingResult);
-    if (!IsActionInterruptable())
-        return;
-    if (SpriteType == PeepSpriteType::Normal)
-    {
-        if (TimeInQueue >= 2000 && (0xFFFF & scenario_rand()) <= 119)
-        {
-            // Eat Food/Look at watch
-            Action = PeepActionType::EatFood;
-            ActionFrame = 0;
-            ActionSpriteImageOffset = 0;
-            UpdateCurrentActionSpriteType();
-        }
-        if (TimeInQueue >= 3500 && (0xFFFF & scenario_rand()) <= 93)
-        {
-            // Create the I have been waiting in line ages thought
-            InsertNewThought(PeepThoughtType::QueuingAges, CurrentRide);
-        }
-    }
-    else
-    {
-        if (!(TimeInQueue & 0x3F) && IsActionIdle() && NextActionSpriteType == PeepActionSpriteType::WatchRide)
-        {
-            switch (SpriteType)
-            {
-                case PeepSpriteType::IceCream:
-                case PeepSpriteType::Chips:
-                case PeepSpriteType::Burger:
-                case PeepSpriteType::Drink:
-                case PeepSpriteType::Candyfloss:
-                case PeepSpriteType::Pizza:
-                case PeepSpriteType::Popcorn:
-                case PeepSpriteType::HotDog:
-                case PeepSpriteType::Tentacle:
-                case PeepSpriteType::ToffeeApple:
-                case PeepSpriteType::Doughnut:
-                case PeepSpriteType::Coffee:
-                case PeepSpriteType::Chicken:
-                case PeepSpriteType::Lemonade:
-                case PeepSpriteType::Pretzel:
-                case PeepSpriteType::SuJongkwa:
-                case PeepSpriteType::Juice:
-                case PeepSpriteType::FunnelCake:
-                case PeepSpriteType::Noodles:
-                case PeepSpriteType::Sausage:
-                case PeepSpriteType::Soup:
-                case PeepSpriteType::Sandwich:
-                    // Eat food
-                    Action = PeepActionType::EatFood;
-                    ActionFrame = 0;
-                    ActionSpriteImageOffset = 0;
-                    UpdateCurrentActionSpriteType();
-                    break;
-                default:
-                    break;
-            }
-        }
-    }
-    if (TimeInQueue < 4300)
-        return;
-
-    if (Happiness <= 65 && (0xFFFF & scenario_rand()) < 2184)
-    {
-        // Give up queueing for the ride
-        sprite_direction ^= (1 << 4);
-        Invalidate();
-        RemoveFromQueue();
-        SetState(PeepState::One);
-    }
 }
 
 /**
@@ -6176,469 +2963,6 @@ static void peep_update_walking_break_scenery(Guest* peep)
     MapInvalidateTileZoom1({ peep->NextLoc, tileElement->GetBaseZ(), tileElement->GetBaseZ() + 32 });
 
     peep->Angriness = 16;
-}
-
-/**
- * rct2: 0x0069101A
- *
- * @return (CF)
- */
-static bool peep_should_watch_ride(TileElement* tileElement)
-{
-    // Ghosts are purely this-client-side and should not cause any interaction,
-    // as that may lead to a desync.
-    if (network_get_mode() != NETWORK_MODE_NONE)
-    {
-        if (tileElement->IsGhost())
-            return false;
-    }
-
-    auto ride = get_ride(tileElement->AsTrack()->GetRideIndex());
-    if (ride == nullptr || !ride->IsRide())
-    {
-        return false;
-    }
-
-    // This is most likely to have peeps watch new rides
-    if (ride->excitement == RIDE_RATING_UNDEFINED)
-    {
-        return true;
-    }
-
-    if (ride->excitement >= RIDE_RATING(4, 70))
-    {
-        return true;
-    }
-
-    if (ride->intensity >= RIDE_RATING(4, 50))
-    {
-        return true;
-    }
-
-    if (ride->GetRideTypeDescriptor().HasFlag(RIDE_TYPE_FLAG_INTERESTING_TO_LOOK_AT))
-    {
-        if ((scenario_rand() & 0xFFFF) > 0x3333)
-        {
-            return false;
-        }
-    }
-    else if (ride->GetRideTypeDescriptor().HasFlag(RIDE_TYPE_FLAG_SLIGHTLY_INTERESTING_TO_LOOK_AT))
-    {
-        if ((scenario_rand() & 0xFFFF) > 0x1000)
-        {
-            return false;
-        }
-    }
-    else
-    {
-        return false;
-    }
-
-    return true;
-}
-
-bool loc_690FD0(Peep* peep, RideId* rideToView, uint8_t* rideSeatToView, TileElement* tileElement)
-{
-    auto ride = get_ride(tileElement->AsTrack()->GetRideIndex());
-    if (ride == nullptr)
-        return false;
-
-    *rideToView = ride->id;
-    if (ride->excitement == RIDE_RATING_UNDEFINED)
-    {
-        *rideSeatToView = 1;
-        if (ride->status != RideStatus::Open)
-        {
-            if (tileElement->GetClearanceZ() > peep->NextLoc.z + (8 * COORDS_Z_STEP))
-            {
-                *rideSeatToView |= (1 << 1);
-            }
-
-            return true;
-        }
-    }
-    else
-    {
-        *rideSeatToView = 0;
-        if (ride->status == RideStatus::Open && !(ride->lifecycle_flags & RIDE_LIFECYCLE_BROKEN_DOWN))
-        {
-            if (tileElement->GetClearanceZ() > peep->NextLoc.z + (8 * COORDS_Z_STEP))
-            {
-                *rideSeatToView = 0x02;
-            }
-
-            return true;
-        }
-    }
-
-    return false;
-}
-
-/**
- *
- *  rct2: 0x00690B99
- *
- * @param edge (eax)
- * @param peep (esi)
- * @param[out] rideToView (cl)
- * @param[out] rideSeatToView (ch)
- * @return !CF
- */
-static bool peep_find_ride_to_look_at(Peep* peep, uint8_t edge, RideId* rideToView, uint8_t* rideSeatToView)
-{
-    TileElement* tileElement;
-
-    auto surfaceElement = MapGetSurfaceElementAt(peep->NextLoc);
-
-    tileElement = reinterpret_cast<TileElement*>(surfaceElement);
-    if (tileElement == nullptr)
-    {
-        return false;
-    }
-
-    do
-    {
-        // Ghosts are purely this-client-side and should not cause any interaction,
-        // as that may lead to a desync.
-        if (network_get_mode() != NETWORK_MODE_NONE)
-        {
-            if (tileElement->IsGhost())
-                continue;
-        }
-        if (tileElement->GetType() != TileElementType::Wall)
-            continue;
-        if (tileElement->GetDirection() != edge)
-            continue;
-        auto wallEntry = tileElement->AsWall()->GetEntry();
-        if (wallEntry == nullptr || (wallEntry->flags2 & WALL_SCENERY_2_IS_OPAQUE))
-            continue;
-        if (peep->NextLoc.z + (4 * COORDS_Z_STEP) <= tileElement->GetBaseZ())
-            continue;
-        if (peep->NextLoc.z + (1 * COORDS_Z_STEP) >= tileElement->GetClearanceZ())
-            continue;
-
-        return false;
-    } while (!(tileElement++)->IsLastForTile());
-
-    uint16_t x = peep->NextLoc.x + CoordsDirectionDelta[edge].x;
-    uint16_t y = peep->NextLoc.y + CoordsDirectionDelta[edge].y;
-    if (!MapIsLocationValid({ x, y }))
-    {
-        return false;
-    }
-
-    surfaceElement = MapGetSurfaceElementAt(CoordsXY{ x, y });
-
-    tileElement = reinterpret_cast<TileElement*>(surfaceElement);
-    if (tileElement == nullptr)
-    {
-        return false;
-    }
-
-    do
-    {
-        // Ghosts are purely this-client-side and should not cause any interaction,
-        // as that may lead to a desync.
-        if (network_get_mode() != NETWORK_MODE_NONE)
-        {
-            if (tileElement->IsGhost())
-                continue;
-        }
-        if (tileElement->GetType() != TileElementType::Wall)
-            continue;
-        if (DirectionReverse(tileElement->GetDirection()) != edge)
-            continue;
-        auto wallEntry = tileElement->AsWall()->GetEntry();
-        if (wallEntry == nullptr || (wallEntry->flags2 & WALL_SCENERY_2_IS_OPAQUE))
-            continue;
-        // TODO: Check whether this shouldn't be <=, as the other loops use. If so, also extract as loop A.
-        if (peep->NextLoc.z + (4 * COORDS_Z_STEP) >= tileElement->GetBaseZ())
-            continue;
-        if (peep->NextLoc.z + (1 * COORDS_Z_STEP) >= tileElement->GetClearanceZ())
-            continue;
-
-        return false;
-    } while (!(tileElement++)->IsLastForTile());
-
-    // TODO: Extract loop B
-    tileElement = reinterpret_cast<TileElement*>(surfaceElement);
-    do
-    {
-        // Ghosts are purely this-client-side and should not cause any interaction,
-        // as that may lead to a desync.
-        if (network_get_mode() != NETWORK_MODE_NONE)
-        {
-            if (tileElement->IsGhost())
-                continue;
-        }
-
-        if (tileElement->GetClearanceZ() + (1 * COORDS_Z_STEP) < peep->NextLoc.z)
-            continue;
-        if (peep->NextLoc.z + (6 * COORDS_Z_STEP) < tileElement->GetBaseZ())
-            continue;
-
-        if (tileElement->GetType() == TileElementType::Track)
-        {
-            if (peep_should_watch_ride(tileElement))
-            {
-                return loc_690FD0(peep, rideToView, rideSeatToView, tileElement);
-            }
-        }
-
-        if (tileElement->GetType() == TileElementType::LargeScenery)
-        {
-            const auto* sceneryEntry = tileElement->AsLargeScenery()->GetEntry();
-            if (sceneryEntry == nullptr || !(sceneryEntry->flags & LARGE_SCENERY_FLAG_PHOTOGENIC))
-            {
-                continue;
-            }
-
-            *rideSeatToView = 0;
-            if (tileElement->GetClearanceZ() >= peep->NextLoc.z + (8 * COORDS_Z_STEP))
-            {
-                *rideSeatToView = 0x02;
-            }
-
-            *rideToView = RideId::GetNull();
-
-            return true;
-        }
-    } while (!(tileElement++)->IsLastForTile());
-
-    // TODO: Extract loop C
-    tileElement = reinterpret_cast<TileElement*>(surfaceElement);
-    do
-    {
-        // Ghosts are purely this-client-side and should not cause any interaction,
-        // as that may lead to a desync.
-        if (network_get_mode() != NETWORK_MODE_NONE)
-        {
-            if (tileElement->IsGhost())
-                continue;
-        }
-        if (tileElement->GetClearanceZ() + (1 * COORDS_Z_STEP) < peep->NextLoc.z)
-            continue;
-        if (peep->NextLoc.z + (6 * COORDS_Z_STEP) < tileElement->GetBaseZ())
-            continue;
-        if (tileElement->GetType() == TileElementType::Surface)
-            continue;
-        if (tileElement->GetType() == TileElementType::Path)
-            continue;
-
-        if (tileElement->GetType() == TileElementType::Wall)
-        {
-            auto wallEntry = tileElement->AsWall()->GetEntry();
-            if (wallEntry == nullptr || (wallEntry->flags2 & WALL_SCENERY_2_IS_OPAQUE))
-            {
-                continue;
-            }
-        }
-
-        return false;
-    } while (!(tileElement++)->IsLastForTile());
-
-    x += CoordsDirectionDelta[edge].x;
-    y += CoordsDirectionDelta[edge].y;
-    if (!MapIsLocationValid({ x, y }))
-    {
-        return false;
-    }
-
-    surfaceElement = MapGetSurfaceElementAt(CoordsXY{ x, y });
-
-    // TODO: extract loop A
-    tileElement = reinterpret_cast<TileElement*>(surfaceElement);
-
-    if (tileElement == nullptr)
-    {
-        return false;
-    }
-
-    do
-    {
-        // Ghosts are purely this-client-side and should not cause any interaction,
-        // as that may lead to a desync.
-        if (network_get_mode() != NETWORK_MODE_NONE)
-        {
-            if (tileElement->IsGhost())
-                continue;
-        }
-        if (tileElement->GetType() != TileElementType::Wall)
-            continue;
-        if (DirectionReverse(tileElement->GetDirection()) != edge)
-            continue;
-        auto wallEntry = tileElement->AsWall()->GetEntry();
-        if (wallEntry == nullptr || (wallEntry->flags2 & WALL_SCENERY_2_IS_OPAQUE))
-            continue;
-        if (peep->NextLoc.z + (6 * COORDS_Z_STEP) <= tileElement->GetBaseZ())
-            continue;
-        if (peep->NextLoc.z >= tileElement->GetClearanceZ())
-            continue;
-
-        return false;
-    } while (!(tileElement++)->IsLastForTile());
-
-    // TODO: Extract loop B
-    tileElement = reinterpret_cast<TileElement*>(surfaceElement);
-    do
-    {
-        // Ghosts are purely this-client-side and should not cause any interaction,
-        // as that may lead to a desync.
-        if (network_get_mode() != NETWORK_MODE_NONE)
-        {
-            if (tileElement->IsGhost())
-                continue;
-        }
-        if (tileElement->GetClearanceZ() + (1 * COORDS_Z_STEP) < peep->NextLoc.z)
-            continue;
-        if (peep->NextLoc.z + (8 * COORDS_Z_STEP) < tileElement->GetBaseZ())
-            continue;
-
-        if (tileElement->GetType() == TileElementType::Track)
-        {
-            if (peep_should_watch_ride(tileElement))
-            {
-                return loc_690FD0(peep, rideToView, rideSeatToView, tileElement);
-            }
-        }
-
-        if (tileElement->GetType() == TileElementType::LargeScenery)
-        {
-            auto* sceneryEntry = tileElement->AsLargeScenery()->GetEntry();
-            if (!(sceneryEntry == nullptr || sceneryEntry->flags & LARGE_SCENERY_FLAG_PHOTOGENIC))
-            {
-                continue;
-            }
-
-            *rideSeatToView = 0;
-            if (tileElement->GetClearanceZ() >= peep->NextLoc.z + (8 * COORDS_Z_STEP))
-            {
-                *rideSeatToView = 0x02;
-            }
-
-            *rideToView = RideId::GetNull();
-
-            return true;
-        }
-    } while (!(tileElement++)->IsLastForTile());
-
-    // TODO: Extract loop C
-    tileElement = reinterpret_cast<TileElement*>(surfaceElement);
-    do
-    {
-        // Ghosts are purely this-client-side and should not cause any interaction,
-        // as that may lead to a desync.
-        if (network_get_mode() != NETWORK_MODE_NONE)
-        {
-            if (tileElement->IsGhost())
-                continue;
-        }
-        if (tileElement->GetClearanceZ() + (1 * COORDS_Z_STEP) < peep->NextLoc.z)
-            continue;
-        if (peep->NextLoc.z + (8 * COORDS_Z_STEP) < tileElement->GetBaseZ())
-            continue;
-        if (tileElement->GetType() == TileElementType::Surface)
-            continue;
-        if (tileElement->GetType() == TileElementType::Path)
-            continue;
-
-        if (tileElement->GetType() == TileElementType::Wall)
-        {
-            auto wallEntry = tileElement->AsWall()->GetEntry();
-            if (wallEntry == nullptr || (wallEntry->flags2 & WALL_SCENERY_2_IS_OPAQUE))
-            {
-                continue;
-            }
-        }
-
-        return false;
-    } while (!(tileElement++)->IsLastForTile());
-
-    x += CoordsDirectionDelta[edge].x;
-    y += CoordsDirectionDelta[edge].y;
-    if (!MapIsLocationValid({ x, y }))
-    {
-        return false;
-    }
-
-    surfaceElement = MapGetSurfaceElementAt(CoordsXY{ x, y });
-
-    // TODO: extract loop A
-    tileElement = reinterpret_cast<TileElement*>(surfaceElement);
-    if (tileElement == nullptr)
-    {
-        return false;
-    }
-
-    do
-    {
-        // Ghosts are purely this-client-side and should not cause any interaction,
-        // as that may lead to a desync.
-        if (network_get_mode() != NETWORK_MODE_NONE)
-        {
-            if (tileElement->IsGhost())
-                continue;
-        }
-        if (tileElement->GetType() != TileElementType::Wall)
-            continue;
-        if (DirectionReverse(tileElement->GetDirection()) != edge)
-            continue;
-        auto wallEntry = tileElement->AsWall()->GetEntry();
-        if (wallEntry == nullptr || (wallEntry->flags2 & WALL_SCENERY_2_IS_OPAQUE))
-            continue;
-        if (peep->NextLoc.z + (8 * COORDS_Z_STEP) <= tileElement->GetBaseZ())
-            continue;
-        if (peep->NextLoc.z >= tileElement->GetClearanceZ())
-            continue;
-
-        return false;
-    } while (!(tileElement++)->IsLastForTile());
-
-    // TODO: Extract loop B
-    tileElement = reinterpret_cast<TileElement*>(surfaceElement);
-    do
-    {
-        // Ghosts are purely this-client-side and should not cause any interaction,
-        // as that may lead to a desync.
-        if (network_get_mode() != NETWORK_MODE_NONE)
-        {
-            if (tileElement->IsGhost())
-                continue;
-        }
-        if (tileElement->GetClearanceZ() + (1 * COORDS_Z_STEP) < peep->NextLoc.z)
-            continue;
-        if (peep->NextLoc.z + (10 * COORDS_Z_STEP) < tileElement->GetBaseZ())
-            continue;
-
-        if (tileElement->GetType() == TileElementType::Track)
-        {
-            if (peep_should_watch_ride(tileElement))
-            {
-                return loc_690FD0(peep, rideToView, rideSeatToView, tileElement);
-            }
-        }
-
-        if (tileElement->GetType() == TileElementType::LargeScenery)
-        {
-            const auto* sceneryEntry = tileElement->AsLargeScenery()->GetEntry();
-            if (sceneryEntry == nullptr || !(sceneryEntry->flags & LARGE_SCENERY_FLAG_PHOTOGENIC))
-            {
-                continue;
-            }
-
-            *rideSeatToView = 0;
-            if (tileElement->GetClearanceZ() >= peep->NextLoc.z + (8 * COORDS_Z_STEP))
-            {
-                *rideSeatToView = 0x02;
-            }
-
-            *rideToView = RideId::GetNull();
-
-            return true;
-        }
-    } while (!(tileElement++)->IsLastForTile());
-
-    return false;
 }
 
 /* Part of 0x0069B8CC rct2: 0x0069BC31 */
@@ -7281,120 +3605,6 @@ int32_t get_peep_face_sprite_large(Guest* peep)
     return face_sprite_large[get_face_sprite_offset(peep)];
 }
 
-/**
- *
- *  rct2: 0x00693CBB
- */
-bool Guest::UpdateQueuePosition(PeepActionType previous_action)
-{
-    TimeInQueue++;
-
-    auto* guestNext = GetEntity<Guest>(GuestNextInQueue);
-    if (guestNext == nullptr)
-    {
-        return false;
-    }
-
-    int16_t x_diff = abs(guestNext->x - x);
-    int16_t y_diff = abs(guestNext->y - y);
-    int16_t z_diff = abs(guestNext->z - z);
-
-    if (z_diff > 10)
-        return false;
-
-    if (x_diff < y_diff)
-    {
-        int16_t temp_x = x_diff;
-        x_diff = y_diff;
-        y_diff = temp_x;
-    }
-
-    x_diff += y_diff / 2;
-    if (x_diff > 7)
-    {
-        if (x_diff > 13)
-        {
-            if ((x & 0xFFE0) != (guestNext->x & 0xFFE0) || (y & 0xFFE0) != (guestNext->y & 0xFFE0))
-                return false;
-        }
-
-        if (sprite_direction != guestNext->sprite_direction)
-            return false;
-
-        switch (guestNext->sprite_direction / 8)
-        {
-            case 0:
-                if (x >= guestNext->x)
-                    return false;
-                break;
-            case 1:
-                if (y <= guestNext->y)
-                    return false;
-                break;
-            case 2:
-                if (x <= guestNext->x)
-                    return false;
-                break;
-            case 3:
-                if (y >= guestNext->y)
-                    return false;
-                break;
-        }
-    }
-
-    if (!IsActionInterruptable())
-        UpdateAction();
-
-    if (!IsActionWalking())
-        return true;
-
-    Action = PeepActionType::Idle;
-    NextActionSpriteType = PeepActionSpriteType::WatchRide;
-    if (previous_action != PeepActionType::Idle)
-        Invalidate();
-    return true;
-}
-
-/**
- *
- *  rct2: 0x006966A9
- */
-void Guest::RemoveFromQueue()
-{
-    auto ride = get_ride(CurrentRide);
-    if (ride == nullptr)
-        return;
-
-    auto& station = ride->GetStation(CurrentRideStation);
-    // Make sure we don't underflow, building while paused might reset it to 0 where peeps have
-    // not yet left the queue.
-    if (station.QueueLength > 0)
-    {
-        station.QueueLength--;
-    }
-
-    if (sprite_index == station.LastPeepInQueue)
-    {
-        station.LastPeepInQueue = GuestNextInQueue;
-        return;
-    }
-
-    auto* otherGuest = GetEntity<Guest>(station.LastPeepInQueue);
-    if (otherGuest == nullptr)
-    {
-        log_error("Invalid Guest Queue list!");
-        return;
-    }
-    for (; otherGuest != nullptr; otherGuest = GetEntity<Guest>(otherGuest->GuestNextInQueue))
-    {
-        if (sprite_index == otherGuest->GuestNextInQueue)
-        {
-            otherGuest->GuestNextInQueue = GuestNextInQueue;
-            return;
-        }
-    }
-}
-
 uint64_t Guest::GetItemFlags() const
 {
     return ItemFlags;
@@ -7423,110 +3633,6 @@ void Guest::GiveItem(ShopItem item)
 bool Guest::HasItem(ShopItem peepItem) const
 {
     return GetItemFlags() & EnumToFlag(peepItem);
-}
-
-static bool IsThoughtShopItemRelated(const PeepThoughtType type)
-{
-    switch (type)
-    {
-        case PeepThoughtType::AlreadyGot:
-        case PeepThoughtType::HaventFinished:
-        case PeepThoughtType::CantAffordItem:
-            return true;
-        default:
-            break;
-    }
-    return false;
-}
-
-void Guest::RemoveRideFromMemory(RideId rideId)
-{
-    if (State == PeepState::Watching)
-    {
-        if (CurrentRide == rideId)
-        {
-            CurrentRide = RideId::GetNull();
-            if (TimeToStand >= 50)
-            {
-                // make peep stop watching the ride
-                TimeToStand = 50;
-            }
-        }
-    }
-
-    // remove any free voucher for this ride from peep
-    if (HasItem(ShopItem::Voucher))
-    {
-        if (VoucherType == VOUCHER_TYPE_RIDE_FREE && VoucherRideId == rideId)
-        {
-            RemoveItem(ShopItem::Voucher);
-        }
-    }
-
-    // remove any photos of this ride from peep
-    if (HasItem(ShopItem::Photo))
-    {
-        if (Photo1RideRef == rideId)
-        {
-            RemoveItem(ShopItem::Photo);
-        }
-    }
-    if (HasItem(ShopItem::Photo2))
-    {
-        if (Photo2RideRef == rideId)
-        {
-            RemoveItem(ShopItem::Photo2);
-        }
-    }
-    if (HasItem(ShopItem::Photo3))
-    {
-        if (Photo3RideRef == rideId)
-        {
-            RemoveItem(ShopItem::Photo3);
-        }
-    }
-    if (HasItem(ShopItem::Photo4))
-    {
-        if (Photo4RideRef == rideId)
-        {
-            RemoveItem(ShopItem::Photo4);
-        }
-    }
-
-    if (GuestHeadingToRideId == rideId)
-    {
-        GuestHeadingToRideId = RideId::GetNull();
-    }
-    if (FavouriteRide == rideId)
-    {
-        FavouriteRide = RideId::GetNull();
-    }
-
-    // Erase all thoughts that contain the ride.
-    for (auto it = std::begin(Thoughts); it != std::end(Thoughts);)
-    {
-        const auto& entry = *it;
-        if (entry.type == PeepThoughtType::None)
-            break;
-
-        // Ride ids and shop item ids might have the same value, look only for ride thoughts.
-        if (IsThoughtShopItemRelated(entry.type) || entry.rideId != rideId)
-        {
-            it++;
-            continue;
-        }
-
-        if (auto itNext = std::next(it); itNext != std::end(Thoughts))
-        {
-            // Overwrite this entry by shifting all entries that follow.
-            std::rotate(it, itNext, std::end(Thoughts));
-        }
-
-        // Last slot is now free.
-        auto& lastEntry = Thoughts.back();
-        lastEntry.type = PeepThoughtType::None;
-        lastEntry.item = PeepThoughtItemNone;
-    }
 }
 
 void Guest::Serialise(DataSerialiser& stream)
@@ -7583,4 +3689,18 @@ void Guest::Serialise(DataSerialiser& stream)
     stream << FavouriteRide;
     stream << FavouriteRideRating;
     stream << ItemFlags;
+}
+
+bool IsThoughtShopItemRelated(const PeepThoughtType type)
+{
+    switch (type)
+    {
+        case PeepThoughtType::AlreadyGot:
+        case PeepThoughtType::HaventFinished:
+        case PeepThoughtType::CantAffordItem:
+            return true;
+        default:
+            break;
+    }
+    return false;
 }
