+++ src/openrct2/ride/RideData.h
@@ -23,6 +23,7 @@
 #include "../common.h"
 #include "../core/BitSet.hpp"
 #include "../entity/Guest.h"
+#include "../entity/guest/GuestRideHelper.h"
 #include "../localisation/StringIds.h"
 #include "../sprites.h"
 #include "../util/Util.h"
@@ -159,7 +160,7 @@ struct UpkeepCostsDescriptor
 using RideTrackGroup = OpenRCT2::BitSet<TRACK_GROUP_COUNT>;
 using UpdateRideApproachVehicleWaypointsFunction = void (*)(Guest&, const CoordsXY&, int16_t&);
 using RideMusicUpdateFunction = void (*)(Ride&);
-using PeepUpdateRideLeaveEntranceFunc = void (*)(Guest*, Ride&, CoordsXYZD&);
+using PeepUpdateRideLeaveEntranceFunc = void (*)(Guest&, Ride&, CoordsXYZD&);
 using StartRideMusicFunction = void (*)(const OpenRCT2::RideAudio::ViewportRideMusicInstance&);
 using LightFXAddLightsMagicVehicleFunction = void (*)(const Vehicle* vehicle);
 using RideLocationFunction = CoordsXY (*)(const Vehicle& vehicle, const Ride& ride, const StationIndex& CurrentRideStation);
