+++ src/openrct2/ride/coaster/meta/WaterCoaster.h
@@ -10,6 +10,7 @@
 #pragma once
 
 #include "../../../drawing/LightFX.h"
+#include "../../../entity/guest/GuestRideHelper.h"
 #include "../../../sprites.h"
 #include "../../RideData.h"
 #include "../../ShopItem.h"
