+++ src/openrct2/ride/gentle/meta/Circus.h
@@ -9,6 +9,7 @@
 
 #pragma once
 
+#include "../../../entity/guest/GuestRideHelper.h"
 #include "../../../sprites.h"
 #include "../../RideData.h"
 #include "../../ShopItem.h"
