+++ src/openrct2/ride/gentle/meta/Maze.h
@@ -9,6 +9,7 @@
 
 #pragma once
 
+#include "../../../entity/guest/GuestRideHelper.h"
 #include "../../../sprites.h"
 #include "../../RideAudio.h"
 #include "../../RideData.h"
